@@ -4,13 +4,12 @@
 MCP tools for campaign analysis and correlation.
 """
 
-import asyncio
-import json
-from datetime import datetime, timedelta
-from typing import Any, Dict, List, Optional, Union
+from datetime import datetime
+from typing import Any
+
 import structlog
 
-from .campaign_analyzer import CampaignAnalyzer, Campaign, CampaignEvent, CorrelationMethod
+from .campaign_analyzer import CampaignAnalyzer, CampaignEvent, CorrelationMethod
 from .elasticsearch_client import ElasticsearchClient
 from .user_config import get_user_config
 
@@ -19,8 +18,8 @@ logger = structlog.get_logger(__name__)
 
 class CampaignMCPTools:
     """MCP tools for campaign analysis and correlation."""
-    
-    def __init__(self, es_client: Optional[ElasticsearchClient] = None):
+
+    def __init__(self, es_client: ElasticsearchClient | None = None):
         """Initialize CampaignMCPTools.
 
         Args:
@@ -30,18 +29,18 @@ class CampaignMCPTools:
         self.es_client = es_client or ElasticsearchClient()
         self.campaign_analyzer = CampaignAnalyzer(self.es_client)
         self.user_config = get_user_config()
-    
+
     async def analyze_campaign(
         self,
-        seed_indicators: List[str],
+        seed_indicators: list[str],
         time_range_hours: int = 168,  # 1 week default
-        correlation_methods: Optional[List[str]] = None,
+        correlation_methods: list[str] | None = None,
         min_confidence: float = 0.7,
         include_timeline: bool = True,
-        include_relationships: bool = True
-    ) -> Dict[str, Any]:
+        include_relationships: bool = True,
+    ) -> dict[str, Any]:
         """Analyze attack campaigns from seed indicators.
-        
+
         Args:
             seed_indicators: List of seed indicators (IPs, domains, etc.)
             time_range_hours: Time range to analyze (default: 168 hours = 1 week)
@@ -49,17 +48,19 @@ class CampaignMCPTools:
             min_confidence: Minimum confidence threshold for campaign inclusion
             include_timeline: Whether to include detailed timeline
             include_relationships: Whether to include indicator relationships
-        
+
         Returns:
             Campaign analysis results with metadata
 
         """
-        logger.info("Starting campaign analysis",
-                   seed_indicators=seed_indicators,
-                   time_range_hours=time_range_hours,
-                   correlation_methods=correlation_methods,
-                   min_confidence=min_confidence)
-        
+        logger.info(
+            "Starting campaign analysis",
+            seed_indicators=seed_indicators,
+            time_range_hours=time_range_hours,
+            correlation_methods=correlation_methods,
+            min_confidence=min_confidence,
+        )
+
         try:
             # Convert correlation methods to enum
             correlation_enums = []
@@ -69,34 +70,34 @@ class CampaignMCPTools:
                         correlation_enums.append(CorrelationMethod(method))
                     except ValueError:
                         logger.warning(f"Unknown correlation method: {method}")
-            
+
             # Default correlation methods if none specified
             if not correlation_enums:
                 correlation_enums = [
                     CorrelationMethod.IP_CORRELATION,
                     CorrelationMethod.INFRASTRUCTURE_CORRELATION,
                     CorrelationMethod.TEMPORAL_CORRELATION,
-                    CorrelationMethod.NETWORK_CORRELATION
+                    CorrelationMethod.NETWORK_CORRELATION,
                 ]
-            
+
             # Get seed events from indicators
             seed_events = await self._get_seed_events(seed_indicators, time_range_hours)
-            
+
             if not seed_events:
                 return {
                     "success": False,
                     "error": "No seed events found for the provided indicators",
-                    "campaign_analysis": None
+                    "campaign_analysis": None,
                 }
-            
+
             # Perform campaign correlation
             campaign = await self.campaign_analyzer.correlate_events(
                 seed_events=seed_events,
                 correlation_criteria=correlation_enums,
                 time_window_hours=time_range_hours,
-                min_confidence=min_confidence
+                min_confidence=min_confidence,
             )
-            
+
             # Build response
             result = {
                 "success": True,
@@ -116,18 +117,18 @@ class CampaignMCPTools:
                     "suspected_actor": campaign.suspected_actor,
                     "campaign_name": campaign.campaign_name,
                     "description": campaign.description,
-                    "related_indicators": campaign.related_indicators
-                }
+                    "related_indicators": campaign.related_indicators,
+                },
             }
-            
+
             # Add timeline if requested
             if include_timeline and campaign.events:
                 timeline = await self.campaign_analyzer.build_campaign_timeline(
                     campaign.events,
-                    timeline_granularity="hourly"
+                    timeline_granularity="hourly",
                 )
                 result["campaign_analysis"]["timeline"] = timeline
-            
+
             # Add relationships if requested
             if include_relationships and campaign.relationships:
                 result["campaign_analysis"]["relationships"] = [
@@ -138,73 +139,77 @@ class CampaignMCPTools:
                         "confidence_score": rel.confidence_score,
                         "evidence": rel.evidence,
                         "first_seen": rel.first_seen.isoformat() if rel.first_seen else None,
-                        "last_seen": rel.last_seen.isoformat() if rel.last_seen else None
+                        "last_seen": rel.last_seen.isoformat() if rel.last_seen else None,
                     }
                     for rel in campaign.relationships
                 ]
-            
-            logger.info("Campaign analysis completed",
-                       campaign_id=campaign.campaign_id,
-                       total_events=campaign.total_events,
-                       confidence_score=campaign.confidence_score)
-            
+
+            logger.info(
+                "Campaign analysis completed",
+                campaign_id=campaign.campaign_id,
+                total_events=campaign.total_events,
+                confidence_score=campaign.confidence_score,
+            )
+
             return result
-            
+
         except Exception as e:
             logger.error("Campaign analysis failed", error=str(e))
             return {
                 "success": False,
-                "error": f"Campaign analysis failed: {str(e)}",
-                "campaign_analysis": None
+                "error": f"Campaign analysis failed: {e!s}",
+                "campaign_analysis": None,
             }
-    
+
     async def expand_campaign_indicators(
         self,
         campaign_id: str,
         expansion_depth: int = 3,
         expansion_strategy: str = "comprehensive",
         include_passive_dns: bool = True,
-        include_threat_intel: bool = True
-    ) -> Dict[str, Any]:
+        include_threat_intel: bool = True,
+    ) -> dict[str, Any]:
         """Expand IOCs to find related indicators.
-        
+
         Args:
             campaign_id: Campaign ID to expand
             expansion_depth: Maximum expansion depth
             expansion_strategy: Expansion strategy (comprehensive, infrastructure, temporal)
             include_passive_dns: Whether to include passive DNS data
             include_threat_intel: Whether to include threat intelligence data
-        
+
         Returns:
             Expanded indicators and relationships
 
         """
-        logger.info("Expanding campaign indicators",
-                   campaign_id=campaign_id,
-                   expansion_depth=expansion_depth,
-                   expansion_strategy=expansion_strategy)
-        
+        logger.info(
+            "Expanding campaign indicators",
+            campaign_id=campaign_id,
+            expansion_depth=expansion_depth,
+            expansion_strategy=expansion_strategy,
+        )
+
         try:
             # Get campaign events
             campaign_events = await self._get_campaign_events(campaign_id)
-            
+
             if not campaign_events:
                 return {
                     "success": False,
                     "error": f"Campaign not found: {campaign_id}",
-                    "expanded_indicators": None
+                    "expanded_indicators": None,
                 }
-            
+
             # Extract IOCs from campaign events
             iocs = self._extract_iocs_from_campaign(campaign_events)
-            
+
             # Expand indicators
             relationships = await self.campaign_analyzer.expand_indicators(
                 seed_iocs=iocs,
                 expansion_strategy=expansion_strategy,
-                max_depth=expansion_depth
+                max_depth=expansion_depth,
             )
-            
+
             # Build response
             expanded_indicators = []
             for rel in relationships:
@@ -215,89 +220,91 @@ class CampaignMCPTools:
                     "source_indicator": rel.source_indicator,
                     "evidence": rel.evidence,
                     "first_seen": rel.first_seen.isoformat() if rel.first_seen else None,
-                    "last_seen": rel.last_seen.isoformat() if rel.last_seen else None
+                    "last_seen": rel.last_seen.isoformat() if rel.last_seen else None,
                 }
-                
+
                 # Add passive DNS data if requested
                 if include_passive_dns:
                     passive_dns = await self._get_passive_dns_data(rel.target_indicator)
                     if passive_dns:
                         indicator_info["passive_dns"] = passive_dns
-                
+
                 # Add threat intelligence data if requested
                 if include_threat_intel:
                     threat_intel = await self._get_threat_intel_data(rel.target_indicator)
                     if threat_intel:
                         indicator_info["threat_intelligence"] = threat_intel
-                
+
                 expanded_indicators.append(indicator_info)
-            
-            logger.info("Campaign indicator expansion completed",
-                       campaign_id=campaign_id,
-                       original_iocs=len(iocs),
-                       expanded_indicators=len(expanded_indicators))
-            
+
+            logger.info(
+                "Campaign indicator expansion completed",
+                campaign_id=campaign_id,
+                original_iocs=len(iocs),
+                expanded_indicators=len(expanded_indicators),
+            )
+
             return {
                 "success": True,
                 "campaign_id": campaign_id,
                 "original_indicators": iocs,
                 "expanded_indicators": expanded_indicators,
-                "total_relationships": len(relationships)
+                "total_relationships": len(relationships),
             }
-            
+
         except Exception as e:
             logger.error("Campaign indicator expansion failed", error=str(e))
             return {
                 "success": False,
-                "error": f"Campaign indicator expansion failed: {str(e)}",
-                "expanded_indicators": None
+                "error": f"Campaign indicator expansion failed: {e!s}",
+                "expanded_indicators": None,
             }
-    
+
     async def get_campaign_timeline(
         self,
         campaign_id: str,
         timeline_granularity: str = "hourly",
         include_event_details: bool = True,
-        include_ttp_analysis: bool = True
-    ) -> Dict[str, Any]:
+        include_ttp_analysis: bool = True,
+    ) -> dict[str, Any]:
         """Build detailed attack timelines.
-        
+
         Args:
             campaign_id: Campaign ID to analyze
             timeline_granularity: Timeline granularity (minute, hourly, daily)
             include_event_details: Whether to include detailed event information
             include_ttp_analysis: Whether to include TTP analysis
-        
+
         Returns:
             Detailed campaign timeline
 
         """
-        logger.info("Building campaign timeline",
-                   campaign_id=campaign_id,
-                   granularity=timeline_granularity)
-        
+        logger.info(
+            "Building campaign timeline", campaign_id=campaign_id, granularity=timeline_granularity
+        )
+
         try:
             # Get campaign events
             campaign_events = await self._get_campaign_events(campaign_id)
-            
+
             if not campaign_events:
                 return {
                     "success": False,
                     "error": f"Campaign not found: {campaign_id}",
-                    "timeline": None
+                    "timeline": None,
                 }
-            
+
             # Build timeline
             timeline_data = await self.campaign_analyzer.build_campaign_timeline(
                 campaign_events,
-                timeline_granularity=timeline_granularity
+                timeline_granularity=timeline_granularity,
             )
-            
+
             # Add TTP analysis if requested
             if include_ttp_analysis:
                 ttp_analysis = self._analyze_campaign_ttps(campaign_events)
                 timeline_data["ttp_analysis"] = ttp_analysis
-            
+
             # Add event details if requested
             if include_event_details:
                 timeline_data["event_details"] = [
@@ -311,51 +318,53 @@ class CampaignMCPTools:
                         "ttp_tactic": event.ttp_tactic,
                         "user_agent": event.user_agent,
                         "url": event.url,
-                        "confidence_score": event.confidence_score
+                        "confidence_score": event.confidence_score,
                     }
                     for event in campaign_events
                 ]
-            
-            logger.info("Campaign timeline built",
-                       campaign_id=campaign_id,
-                       timeline_periods=timeline_data["total_periods"],
-                       total_events=timeline_data["total_events"])
-            
+
+            logger.info(
+                "Campaign timeline built",
+                campaign_id=campaign_id,
+                timeline_periods=timeline_data["total_periods"],
+                total_events=timeline_data["total_events"],
+            )
+
             return {
                 "success": True,
                 "campaign_id": campaign_id,
-                "timeline": timeline_data
+                "timeline": timeline_data,
             }
-            
+
         except Exception as e:
             logger.error("Campaign timeline build failed", error=str(e))
             return {
                 "success": False,
-                "error": f"Campaign timeline build failed: {str(e)}",
-                "timeline": None
+                "error": f"Campaign timeline build failed: {e!s}",
+                "timeline": None,
             }
-    
+
     async def compare_campaigns(
         self,
-        campaign_ids: List[str],
-        comparison_metrics: Optional[List[str]] = None,
-        include_visualization_data: bool = True
-    ) -> Dict[str, Any]:
+        campaign_ids: list[str],
+        comparison_metrics: list[str] | None = None,
+        include_visualization_data: bool = True,
+    ) -> dict[str, Any]:
         """Compare multiple campaigns for similarities.
-        
+
         Args:
             campaign_ids: List of campaign IDs to compare
             comparison_metrics: Metrics to compare (ttps, infrastructure, timing, etc.)
             include_visualization_data: Whether to include visualization data
-        
+
         Returns:
             Campaign comparison results
 
         """
-        logger.info("Comparing campaigns",
-                   campaign_ids=campaign_ids,
-                   comparison_metrics=comparison_metrics)
-        
+        logger.info(
+            "Comparing campaigns", campaign_ids=campaign_ids, comparison_metrics=comparison_metrics
+        )
+
         try:
             # Get campaign data
             campaigns = []
@@ -363,235 +372,260 @@ class CampaignMCPTools:
                 campaign_data = await self._get_campaign_data(campaign_id)
                 if campaign_data:
                     campaigns.append(campaign_data)
-            
+
             if len(campaigns) < 2:
                 return {
                     "success": False,
                     "error": "At least 2 campaigns required for comparison",
-                    "comparison": None
+                    "comparison": None,
                 }
-            
+
             # Default comparison metrics
             if not comparison_metrics:
                 comparison_metrics = [
-                    "ttps", "infrastructure", "timing", "geography", "sophistication"
+                    "ttps",
+                    "infrastructure",
+                    "timing",
+                    "geography",
+                    "sophistication",
                 ]
-            
+
             # Perform comparisons
             comparison_results = {}
-            
+
             if "ttps" in comparison_metrics:
                 comparison_results["ttp_similarity"] = self._compare_campaign_ttps(campaigns)
-            
+
             if "infrastructure" in comparison_metrics:
-                comparison_results["infrastructure_similarity"] = self._compare_campaign_infrastructure(campaigns)
-            
+                comparison_results["infrastructure_similarity"] = (
+                    self._compare_campaign_infrastructure(campaigns)
+                )
+
             if "timing" in comparison_metrics:
                 comparison_results["timing_similarity"] = self._compare_campaign_timing(campaigns)
-            
+
             if "geography" in comparison_metrics:
-                comparison_results["geographic_similarity"] = self._compare_campaign_geography(campaigns)
-            
+                comparison_results["geographic_similarity"] = self._compare_campaign_geography(
+                    campaigns
+                )
+
             if "sophistication" in comparison_metrics:
-                comparison_results["sophistication_comparison"] = self._compare_campaign_sophistication(campaigns)
-            
+                comparison_results["sophistication_comparison"] = (
+                    self._compare_campaign_sophistication(campaigns)
+                )
+
             # Calculate overall similarity scores
             similarity_matrix = self._calculate_similarity_matrix(campaigns, comparison_results)
-            
+
             result = {
                 "success": True,
                 "campaign_ids": campaign_ids,
                 "comparison_metrics": comparison_metrics,
                 "similarity_matrix": similarity_matrix,
-                "detailed_comparisons": comparison_results
+                "detailed_comparisons": comparison_results,
             }
-            
+
             # Add visualization data if requested
             if include_visualization_data:
                 result["visualization_data"] = self._generate_comparison_visualization_data(
-                    campaigns, comparison_results
+                    campaigns,
+                    comparison_results,
                 )
-            
-            logger.info("Campaign comparison completed",
-                       campaign_count=len(campaigns),
-                       comparison_metrics=comparison_metrics)
-            
+
+            logger.info(
+                "Campaign comparison completed",
+                campaign_count=len(campaigns),
+                comparison_metrics=comparison_metrics,
+            )
+
             return result
-            
+
         except Exception as e:
             logger.error("Campaign comparison failed", error=str(e))
             return {
                 "success": False,
-                "error": f"Campaign comparison failed: {str(e)}",
-                "comparison": None
+                "error": f"Campaign comparison failed: {e!s}",
+                "comparison": None,
             }
-    
+
     async def detect_ongoing_campaigns(
         self,
         time_window_hours: int = 24,
         min_event_threshold: int = 15,
         correlation_threshold: float = 0.8,
-        include_alert_data: bool = True
-    ) -> Dict[str, Any]:
+        include_alert_data: bool = True,
+    ) -> dict[str, Any]:
         """Real-time detection of active campaigns.
-        
+
         Args:
             time_window_hours: Time window for detection (default: 24 hours)
             min_event_threshold: Minimum events for campaign detection
             correlation_threshold: Minimum correlation threshold
             include_alert_data: Whether to include alert data
-        
+
         Returns:
             Detected ongoing campaigns
 
         """
-        logger.info("Detecting ongoing campaigns",
-                   time_window_hours=time_window_hours,
-                   min_event_threshold=min_event_threshold,
-                   correlation_threshold=correlation_threshold)
-        
+        logger.info(
+            "Detecting ongoing campaigns",
+            time_window_hours=time_window_hours,
+            min_event_threshold=min_event_threshold,
+            correlation_threshold=correlation_threshold,
+        )
+
         try:
             # Query recent events
             recent_events, _, _ = await self.es_client.query_dshield_events(
                 time_range_hours=time_window_hours,
-                page_size=1000
+                page_size=1000,
             )
-            
+
             if len(recent_events) < min_event_threshold:
                 return {
                     "success": True,
                     "ongoing_campaigns": [],
                     "total_events_analyzed": len(recent_events),
-                    "message": "Insufficient events for campaign detection"
+                    "message": "Insufficient events for campaign detection",
                 }
-            
+
             # Group events by potential campaigns
             campaign_groups = self._group_events_by_campaigns(recent_events, correlation_threshold)
-            
+
             # Analyze each group
             ongoing_campaigns = []
             for group_id, events in campaign_groups.items():
                 if len(events) >= min_event_threshold:
                     campaign_analysis = await self._analyze_campaign_group(
-                        group_id, events, include_alert_data
+                        group_id,
+                        events,
+                        include_alert_data,
                     )
                     if campaign_analysis:
                         ongoing_campaigns.append(campaign_analysis)
-            
+
             # Sort by threat level
             ongoing_campaigns.sort(key=lambda x: x.get("threat_level_score", 0), reverse=True)
-            
-            logger.info("Ongoing campaign detection completed",
-                       total_events=len(recent_events),
-                       campaign_groups=len(campaign_groups),
-                       ongoing_campaigns=len(ongoing_campaigns))
-            
+
+            logger.info(
+                "Ongoing campaign detection completed",
+                total_events=len(recent_events),
+                campaign_groups=len(campaign_groups),
+                ongoing_campaigns=len(ongoing_campaigns),
+            )
+
             return {
                 "success": True,
                 "ongoing_campaigns": ongoing_campaigns,
                 "total_events_analyzed": len(recent_events),
                 "detection_time": datetime.now().isoformat(),
-                "time_window_hours": time_window_hours
+                "time_window_hours": time_window_hours,
             }
-            
+
         except Exception as e:
             logger.error("Ongoing campaign detection failed", error=str(e))
             return {
                 "success": False,
-                "error": f"Ongoing campaign detection failed: {str(e)}",
-                "ongoing_campaigns": None
+                "error": f"Ongoing campaign detection failed: {e!s}",
+                "ongoing_campaigns": None,
             }
-    
+
     async def search_campaigns(
         self,
-        search_criteria: Dict[str, Any],
+        search_criteria: dict[str, Any],
         time_range_hours: int = 168,  # 1 week
         max_results: int = 50,
-        include_summaries: bool = True
-    ) -> Dict[str, Any]:
+        include_summaries: bool = True,
+    ) -> dict[str, Any]:
         """Search existing campaigns by criteria.
-        
+
         Args:
             search_criteria: Search criteria (indicators, time_range, confidence, etc.)
             time_range_hours: Time range for search
             max_results: Maximum results to return
             include_summaries: Whether to include campaign summaries
-        
+
         Returns:
             Matching campaigns
 
         """
-        logger.info("Searching campaigns",
-                   search_criteria=search_criteria,
-                   time_range_hours=time_range_hours,
-                   max_results=max_results)
-        
+        logger.info(
+            "Searching campaigns",
+            search_criteria=search_criteria,
+            time_range_hours=time_range_hours,
+            max_results=max_results,
+        )
+
         try:
             # Build search query
             search_query = self._build_campaign_search_query(search_criteria, time_range_hours)
-            
+
             # Search for campaigns
             matching_campaigns = await self._search_campaign_database(search_query, max_results)
-            
+
             # Add summaries if requested
             if include_summaries:
                 for campaign in matching_campaigns:
                     campaign["summary"] = self._generate_campaign_summary(campaign)
-            
-            logger.info("Campaign search completed",
-                       search_criteria=search_criteria,
-                       matching_campaigns=len(matching_campaigns))
-            
+
+            logger.info(
+                "Campaign search completed",
+                search_criteria=search_criteria,
+                matching_campaigns=len(matching_campaigns),
+            )
+
             return {
                 "success": True,
                 "search_criteria": search_criteria,
                 "matching_campaigns": matching_campaigns,
                 "total_results": len(matching_campaigns),
-                "search_time": datetime.now().isoformat()
+                "search_time": datetime.now().isoformat(),
             }
-            
+
         except Exception as e:
             logger.error("Campaign search failed", error=str(e))
             return {
                 "success": False,
-                "error": f"Campaign search failed: {str(e)}",
-                "matching_campaigns": None
+                "error": f"Campaign search failed: {e!s}",
+                "matching_campaigns": None,
             }
-    
+
     async def get_campaign_details(
         self,
         campaign_id: str,
         include_full_timeline: bool = False,
         include_relationships: bool = True,
-        include_threat_intel: bool = True
-    ) -> Dict[str, Any]:
+        include_threat_intel: bool = True,
+    ) -> dict[str, Any]:
         """Comprehensive campaign information.
-        
+
         Args:
             campaign_id: Campaign ID to retrieve
             include_full_timeline: Whether to include full timeline
             include_relationships: Whether to include indicator relationships
             include_threat_intel: Whether to include threat intelligence
-        
+
         Returns:
             Comprehensive campaign details
 
         """
-        logger.info("Getting campaign details",
-                   campaign_id=campaign_id,
-                   include_full_timeline=include_full_timeline)
-        
+        logger.info(
+            "Getting campaign details",
+            campaign_id=campaign_id,
+            include_full_timeline=include_full_timeline,
+        )
+
         try:
             # Get campaign data
             campaign_data = await self._get_campaign_data(campaign_id)
-            
+
             if not campaign_data:
                 return {
                     "success": False,
                     "error": f"Campaign not found: {campaign_id}",
-                    "campaign_details": None
+                    "campaign_details": None,
                 }
-            
+
             # Build comprehensive details
             details = {
                 "campaign_id": campaign_id,
@@ -604,32 +638,32 @@ class CampaignMCPTools:
                     "unique_targets": campaign_data.get("unique_targets"),
                     "suspected_actor": campaign_data.get("suspected_actor"),
                     "campaign_name": campaign_data.get("campaign_name"),
-                    "description": campaign_data.get("description")
+                    "description": campaign_data.get("description"),
                 },
                 "attack_analysis": {
                     "attack_vectors": campaign_data.get("attack_vectors", []),
                     "ttp_techniques": campaign_data.get("ttp_techniques", []),
                     "ttp_tactics": campaign_data.get("ttp_tactics", []),
-                    "sophistication_score": self._calculate_sophistication_score(campaign_data)
+                    "sophistication_score": self._calculate_sophistication_score(campaign_data),
                 },
                 "infrastructure": {
                     "domains": campaign_data.get("infrastructure_domains", []),
                     "ips": campaign_data.get("related_indicators", []),
-                    "geographic_regions": campaign_data.get("geographic_regions", [])
-                }
+                    "geographic_regions": campaign_data.get("geographic_regions", []),
+                },
             }
-            
+
             # Add full timeline if requested
             if include_full_timeline:
                 timeline = await self.get_campaign_timeline(
                     campaign_id,
                     timeline_granularity="hourly",
                     include_event_details=True,
-                    include_ttp_analysis=True
+                    include_ttp_analysis=True,
                 )
                 if timeline.get("success"):
                     details["timeline"] = timeline["timeline"]
-            
+
             # Add relationships if requested
             if include_relationships:
                 relationships = await self.expand_campaign_indicators(
@@ -637,116 +671,147 @@ class CampaignMCPTools:
                     expansion_depth=2,
                     expansion_strategy="comprehensive",
                     include_passive_dns=True,
-                    include_threat_intel=include_threat_intel
+                    include_threat_intel=include_threat_intel,
                 )
                 if relationships.get("success"):
                     details["relationships"] = relationships["expanded_indicators"]
-            
+
             # Add threat intelligence if requested
             if include_threat_intel:
                 threat_intel = await self._get_campaign_threat_intel(campaign_id)
                 if threat_intel:
                     details["threat_intelligence"] = threat_intel
-            
-            logger.info("Campaign details retrieved",
-                       campaign_id=campaign_id,
-                       include_full_timeline=include_full_timeline)
-            
+
+            logger.info(
+                "Campaign details retrieved",
+                campaign_id=campaign_id,
+                include_full_timeline=include_full_timeline,
+            )
+
             return {
                 "success": True,
-                "campaign_details": details
+                "campaign_details": details,
             }
-            
+
         except Exception as e:
             logger.error("Campaign details retrieval failed", error=str(e))
             return {
                 "success": False,
-                "error": f"Campaign details retrieval failed: {str(e)}",
-                "campaign_details": None
+                "error": f"Campaign details retrieval failed: {e!s}",
+                "campaign_details": None,
             }
-    
+
     # Helper methods
-    
-    async def _get_seed_events(self, indicators: List[str], time_range_hours: int) -> List[Dict[str, Any]]:
+
+    async def _get_seed_events(
+        self, indicators: list[str], time_range_hours: int
+    ) -> list[dict[str, Any]]:
         """Get seed events from indicators."""
         all_events = []
         # Dynamically get all possible IP fields from the ElasticsearchClient field mappings
         ip_fields = []
         try:
-            ip_fields = ElasticsearchClient().dshield_field_mappings['source_ip'] + ElasticsearchClient().dshield_field_mappings['destination_ip']
+            ip_fields = (
+                ElasticsearchClient().dshield_field_mappings["source_ip"]
+                + ElasticsearchClient().dshield_field_mappings["destination_ip"]
+            )
         except Exception as e:
             logger.warning(f"Could not load IP field mappings: {e}")
             ip_fields = ["source.ip", "destination.ip", "related.ip"]
-        
+
         for indicator in indicators:
             try:
-                # Try each IP field individually with a simple filter (like the working debug script)
+                # Try each IP field individually with a simple filter
+                # (like the working debug script)
                 for field in ip_fields:
                     try:
                         # Use simple filter approach that we know works
                         filters = {field: indicator}
-                        
+
                         # DEBUG: Log the filters being sent
                         import json
-                        logger.debug(f"_get_seed_events: Querying with simple filter for indicator {indicator}, field {field}: {json.dumps(filters, indent=2)}")
-                        
+
+                        logger.debug(
+                            f"_get_seed_events: Querying with simple filter for "
+                            f"indicator {indicator}, field {field}: "
+                            f"{json.dumps(filters, indent=2)}"
+                        )
+
                         events, _, _ = await self.es_client.query_dshield_events(
                             time_range_hours=time_range_hours,
                             filters=filters,
-                            page_size=100
+                            page_size=100,
                         )
-                        
+
                         if events:
-                            logger.debug(f"_get_seed_events: Found {len(events)} events for indicator {indicator} in field {field}")
+                            logger.debug(
+                                f"_get_seed_events: Found {len(events)} events for "
+                                f"indicator {indicator} in field {field}"
+                            )
                             all_events.extend(events)
-                            # If we found events for this field, no need to try other fields for this indicator
+                            # If we found events for this field, no need to try other
+                            # fields for this indicator
                             break
-                        else:
-                            logger.debug(f"_get_seed_events: No events found for indicator {indicator} in field {field}")
-                            
+                        logger.debug(
+                            f"_get_seed_events: No events found for indicator "
+                            f"{indicator} in field {field}"
+                        )
+
                     except Exception as field_error:
-                        logger.warning(f"Failed to query field {field} for indicator {indicator}: {field_error}")
+                        logger.warning(
+                            f"Failed to query field {field} for indicator "
+                            f"{indicator}: {field_error}"
+                        )
                         continue
-                
+
                 # Also try URL and user agent wildcards if no IP events found
-                if not any(event for event in all_events if event.get('source_ip') == indicator or event.get('destination_ip') == indicator):
+                if not any(
+                    event
+                    for event in all_events
+                    if event.get("source_ip") == indicator
+                    or event.get("destination_ip") == indicator
+                ):
                     try:
                         # Try URL wildcard
                         url_filters = {"url.original": f"*{indicator}*"}
                         events, _, _ = await self.es_client.query_dshield_events(
                             time_range_hours=time_range_hours,
                             filters=url_filters,
-                            page_size=100
+                            page_size=100,
                         )
                         if events:
                             all_events.extend(events)
                             continue
                     except Exception as url_error:
-                        logger.warning(f"Failed to query URL for indicator {indicator}: {url_error}")
-                    
+                        logger.warning(
+                            f"Failed to query URL for indicator {indicator}: {url_error}"
+                        )
+
                     try:
                         # Try user agent wildcard
                         ua_filters = {"user_agent.original": f"*{indicator}*"}
                         events, _, _ = await self.es_client.query_dshield_events(
                             time_range_hours=time_range_hours,
                             filters=ua_filters,
-                            page_size=100
+                            page_size=100,
                         )
                         if events:
                             all_events.extend(events)
                     except Exception as ua_error:
-                        logger.warning(f"Failed to query user agent for indicator {indicator}: {ua_error}")
-                        
+                        logger.warning(
+                            f"Failed to query user agent for indicator {indicator}: {ua_error}"
+                        )
+
             except Exception as e:
                 logger.warning(f"Failed to get events for indicator {indicator}: {e}")
-        
+
         logger.info(f"_get_seed_events: Total events found for all indicators: {len(all_events)}")
         return all_events
-    
-    def _extract_iocs_from_campaign(self, campaign_events: List[CampaignEvent]) -> List[str]:
+
+    def _extract_iocs_from_campaign(self, campaign_events: list[CampaignEvent]) -> list[str]:
         """Extract IOCs from campaign events."""
         iocs = set()
-        
+
         for event in campaign_events:
             if event.source_ip:
                 iocs.add(event.source_ip)
@@ -757,20 +822,21 @@ class CampaignMCPTools:
                 domain = self._extract_domain_from_url(event.url)
                 if domain:
                     iocs.add(domain)
-        
+
         return list(iocs)
-    
-    def _extract_domain_from_url(self, url: str) -> Optional[str]:
+
+    def _extract_domain_from_url(self, url: str) -> str | None:
         """Extract domain from URL."""
         try:
             import re
-            domain_pattern = r'https?://([^/]+)'
+
+            domain_pattern = r"https?://([^/]+)"
             match = re.search(domain_pattern, url)
             return match.group(1) if match else None
-        except:
+        except Exception:
             return None
-    
-    async def _get_campaign_events(self, campaign_id: str) -> List[CampaignEvent]:
+
+    async def _get_campaign_events(self, campaign_id: str) -> list[CampaignEvent]:
         """Get campaign events by campaign ID."""
         # This is a simplified implementation
         # In a real system, you'd query a campaign database
@@ -781,8 +847,8 @@ class CampaignMCPTools:
         except Exception as e:
             logger.error(f"Failed to get campaign events for {campaign_id}: {e}")
             return []
-    
-    async def _get_campaign_data(self, campaign_id: str) -> Optional[Dict[str, Any]]:
+
+    async def _get_campaign_data(self, campaign_id: str) -> dict[str, Any] | None:
         """Get campaign data by campaign ID."""
         # This is a simplified implementation
         # In a real system, you'd query a campaign database
@@ -793,106 +859,118 @@ class CampaignMCPTools:
         except Exception as e:
             logger.error(f"Failed to get campaign data for {campaign_id}: {e}")
             return None
-    
-    async def _get_passive_dns_data(self, indicator: str) -> Optional[Dict[str, Any]]:
+
+    async def _get_passive_dns_data(self, indicator: str) -> dict[str, Any] | None:
         """Get passive DNS data for an indicator."""
         # This is a simplified implementation
         # In a real system, you'd integrate with passive DNS services
         return None
-    
-    async def _get_threat_intel_data(self, indicator: str) -> Optional[Dict[str, Any]]:
+
+    async def _get_threat_intel_data(self, indicator: str) -> dict[str, Any] | None:
         """Get threat intelligence data for an indicator."""
         # This is a simplified implementation
         # In a real system, you'd integrate with threat intelligence feeds
         return None
-    
-    def _analyze_campaign_ttps(self, campaign_events: List[CampaignEvent]) -> Dict[str, Any]:
+
+    def _analyze_campaign_ttps(self, campaign_events: list[CampaignEvent]) -> dict[str, Any]:
         """Analyze TTPs in campaign events."""
-        ttp_counts = {}
-        tactic_counts = {}
-        
+        ttp_counts: dict[str, int] = {}
+        tactic_counts: dict[str, int] = {}
+
         for event in campaign_events:
             if event.ttp_technique:
                 ttp_counts[event.ttp_technique] = ttp_counts.get(event.ttp_technique, 0) + 1
             if event.ttp_tactic:
                 tactic_counts[event.ttp_tactic] = tactic_counts.get(event.ttp_tactic, 0) + 1
-        
+
         return {
             "techniques": ttp_counts,
             "tactics": tactic_counts,
             "total_techniques": len(ttp_counts),
-            "total_tactics": len(tactic_counts)
+            "total_tactics": len(tactic_counts),
         }
-    
-    def _compare_campaign_ttps(self, campaigns: List[Dict[str, Any]]) -> Dict[str, Any]:
+
+    def _compare_campaign_ttps(self, campaigns: list[dict[str, Any]]) -> dict[str, Any]:
         """Compare TTPs across campaigns."""
         # Simplified implementation
         return {"similarity_score": 0.5}
-    
-    def _compare_campaign_infrastructure(self, campaigns: List[Dict[str, Any]]) -> Dict[str, Any]:
+
+    def _compare_campaign_infrastructure(self, campaigns: list[dict[str, Any]]) -> dict[str, Any]:
         """Compare infrastructure across campaigns."""
         # Simplified implementation
         return {"similarity_score": 0.5}
-    
-    def _compare_campaign_timing(self, campaigns: List[Dict[str, Any]]) -> Dict[str, Any]:
+
+    def _compare_campaign_timing(self, campaigns: list[dict[str, Any]]) -> dict[str, Any]:
         """Compare timing patterns across campaigns."""
         # Simplified implementation
         return {"similarity_score": 0.5}
-    
-    def _compare_campaign_geography(self, campaigns: List[Dict[str, Any]]) -> Dict[str, Any]:
+
+    def _compare_campaign_geography(self, campaigns: list[dict[str, Any]]) -> dict[str, Any]:
         """Compare geographic patterns across campaigns."""
         # Simplified implementation
         return {"similarity_score": 0.5}
-    
-    def _compare_campaign_sophistication(self, campaigns: List[Dict[str, Any]]) -> Dict[str, Any]:
+
+    def _compare_campaign_sophistication(self, campaigns: list[dict[str, Any]]) -> dict[str, Any]:
         """Compare sophistication levels across campaigns."""
         # Simplified implementation
         return {"similarity_score": 0.5}
-    
-    def _calculate_similarity_matrix(self, campaigns: List[Dict[str, Any]], comparison_results: Dict[str, Any]) -> List[List[float]]:
+
+    def _calculate_similarity_matrix(
+        self, campaigns: list[dict[str, Any]], comparison_results: dict[str, Any]
+    ) -> list[list[float]]:
         """Calculate similarity matrix for campaigns."""
         # Simplified implementation
         return [[1.0, 0.5], [0.5, 1.0]]
-    
-    def _generate_comparison_visualization_data(self, campaigns: List[Dict[str, Any]], comparison_results: Dict[str, Any]) -> Dict[str, Any]:
+
+    def _generate_comparison_visualization_data(
+        self, campaigns: list[dict[str, Any]], comparison_results: dict[str, Any]
+    ) -> dict[str, Any]:
         """Generate visualization data for campaign comparison."""
         # Simplified implementation
         return {"visualization_type": "similarity_matrix"}
-    
-    def _group_events_by_campaigns(self, events: List[Dict[str, Any]], correlation_threshold: float) -> Dict[str, List[Dict[str, Any]]]:
+
+    def _group_events_by_campaigns(
+        self, events: list[dict[str, Any]], correlation_threshold: float
+    ) -> dict[str, list[dict[str, Any]]]:
         """Group events by potential campaigns."""
         # Simplified implementation
-        return {"group_1": events[:len(events)//2], "group_2": events[len(events)//2:]}
-    
-    async def _analyze_campaign_group(self, group_id: str, events: List[Dict[str, Any]], include_alert_data: bool) -> Optional[Dict[str, Any]]:
+        return {"group_1": events[: len(events) // 2], "group_2": events[len(events) // 2 :]}
+
+    async def _analyze_campaign_group(
+        self, group_id: str, events: list[dict[str, Any]], include_alert_data: bool
+    ) -> dict[str, Any] | None:
         """Analyze a group of events as a potential campaign."""
         # Simplified implementation
         return {
             "campaign_id": group_id,
             "threat_level_score": 0.7,
-            "event_count": len(events)
+            "event_count": len(events),
         }
-    
-    def _build_campaign_search_query(self, search_criteria: Dict[str, Any], time_range_hours: int) -> Dict[str, Any]:
+
+    def _build_campaign_search_query(
+        self, search_criteria: dict[str, Any], time_range_hours: int
+    ) -> dict[str, Any]:
         """Build search query for campaigns."""
         # Simplified implementation
         return {"query": "search"}
-    
-    async def _search_campaign_database(self, search_query: Dict[str, Any], max_results: int) -> List[Dict[str, Any]]:
+
+    async def _search_campaign_database(
+        self, search_query: dict[str, Any], max_results: int
+    ) -> list[dict[str, Any]]:
         """Search campaign database."""
         # Simplified implementation
         return []
-    
-    def _generate_campaign_summary(self, campaign: Dict[str, Any]) -> str:
+
+    def _generate_campaign_summary(self, campaign: dict[str, Any]) -> str:
         """Generate campaign summary."""
         return f"Campaign with {campaign.get('total_events', 0)} events"
-    
-    def _calculate_sophistication_score(self, campaign_data: Dict[str, Any]) -> float:
+
+    def _calculate_sophistication_score(self, campaign_data: dict[str, Any]) -> float:
         """Calculate sophistication score for campaign."""
         # Simplified implementation
         return 0.7
-    
-    async def _get_campaign_threat_intel(self, campaign_id: str) -> Optional[Dict[str, Any]]:
+
+    async def _get_campaign_threat_intel(self, campaign_id: str) -> dict[str, Any] | None:
         """Get threat intelligence for campaign."""
         # Simplified implementation
-        return None 
+        return None
