@@ -18,100 +18,124 @@ Example:
     >>> pytest tests/test_enhanced_threat_intelligence.py -v
 """
 
-import pytest
-import pytest_asyncio
 import asyncio
-from datetime import datetime, timedelta
-from typing import Dict, Any
-from unittest.mock import AsyncMock, MagicMock, patch
 import os
 import uuid
+from datetime import datetime, timedelta
+from typing import Any
+from unittest.mock import AsyncMock, MagicMock, Mock, patch
 
+import pytest
+import pytest_asyncio
+
+from src.models import DomainIntelligence, ThreatIntelligenceResult, ThreatIntelligenceSource
 from src.threat_intelligence_manager import ThreatIntelligenceManager
-from src.models import ThreatIntelligenceResult, DomainIntelligence, ThreatIntelligenceSource
 
 
-pytestmark = pytest.mark.asyncio
+# Fixture to mock config functions for all tests
+@pytest.fixture(autouse=True)
+def mock_config_functions():
+    """Mock get_config and get_user_config to prevent subprocess calls."""
+    with (
+        patch('src.threat_intelligence_manager.get_config') as mock_get_config,
+        patch('src.threat_intelligence_manager.get_user_config') as mock_get_user_config,
+        patch('src.threat_intelligence_manager.DShieldClient') as mock_dshield_client,
+    ):
+        mock_config = {"dshield": {"api_key": "test_key"}, "threat_intel": {"cache_ttl_hours": 1}}
+        mock_user_config = Mock()
+        mock_user_config.performance_settings.enable_sqlite_cache = False
+        mock_user_config.performance_settings.sqlite_cache_ttl_hours = 24
+        mock_user_config.get_cache_database_path.return_value = "/tmp/test.db"
+        mock_get_config.return_value = mock_config
+        mock_get_user_config.return_value = mock_user_config
+
+        # Mock DShield client
+        mock_dshield_instance = Mock()
+        mock_dshield_instance.__class__.__name__ = "DShieldClient"
+        mock_dshield_client.return_value = mock_dshield_instance
+
+        yield mock_config, mock_user_config
+
+
+# Individual async tests are marked with @pytest.mark.asyncio
 
 
 class TestThreatIntelligenceManager:
     """Test suite for Threat Intelligence Manager."""
-    
+
     @pytest_asyncio.fixture
     async def threat_manager(self) -> ThreatIntelligenceManager:
         """Create a Threat Intelligence Manager instance for testing."""
         manager = ThreatIntelligenceManager()
         yield manager
         await manager.cleanup()
-        
+
         # Clean up SQLite cache file after each test
         if manager.sqlite_cache_enabled and os.path.exists(manager.sqlite_cache_path):
             try:
                 os.remove(manager.sqlite_cache_path)
             except Exception:
                 pass  # Ignore cleanup errors
-    
+
     @pytest.fixture
-    def mock_config(self) -> Dict[str, Any]:
+    def mock_config(self) -> dict[str, Any]:
         """Mock configuration for testing."""
         return {
             "threat_intelligence": {
                 "sources": {
                     "dshield": {"enabled": True, "rate_limit_requests_per_minute": 60},
                     "virustotal": {"enabled": False},
-                    "shodan": {"enabled": False}
+                    "shodan": {"enabled": False},
                 },
-                "correlation": {
-                    "confidence_threshold": 0.7,
-                    "max_sources_per_query": 3
-                },
+                "correlation": {"confidence_threshold": 0.7, "max_sources_per_query": 3},
                 "cache_ttl_hours": 1,
-                "max_cache_size": 1000
+                "max_cache_size": 1000,
             }
         }
-    
+
     def mock_user_config(self):
         """Create a mock user config for tests that need it."""
         config = MagicMock()
         config.get_setting.return_value = "default_value"
-        
+
         # Mock performance settings for SQLite cache
         performance_settings = MagicMock()
         performance_settings.enable_sqlite_cache = True
         performance_settings.sqlite_cache_ttl_hours = 24
         performance_settings.sqlite_cache_db_name = "test_enrichment_cache.sqlite3"
         config.performance_settings = performance_settings
-        
+
         # Mock database path methods
         config.get_database_directory.return_value = "/tmp/test_db"
         config.get_cache_database_path.return_value = "/tmp/test_db/test_enrichment_cache.sqlite3"
-        
+
         return config
-    
+
     @pytest.mark.asyncio
     async def test_manager_initialization(self, mock_config):
         """Test Threat Intelligence Manager initialization."""
         mock_user_config = self.mock_user_config()
-        with patch('src.threat_intelligence_manager.get_config', return_value=mock_config), \
-             patch('src.threat_intelligence_manager.get_user_config', return_value=mock_user_config):
-            
+        with (
+            patch('src.threat_intelligence_manager.get_config', return_value=mock_config),
+            patch('src.threat_intelligence_manager.get_user_config', return_value=mock_user_config),
+        ):
             manager = ThreatIntelligenceManager()
-            
+
             assert manager.config == mock_config
             assert manager.user_config == mock_user_config
             assert manager.confidence_threshold == 0.7
             assert manager.max_sources == 3
             assert manager.cache_ttl == timedelta(hours=1)
             assert ThreatIntelligenceSource.DSHIELD in manager.clients
-            
+
             await manager.cleanup()
-    
+
     @pytest.mark.asyncio
     async def test_enrich_ip_comprehensive_success(self, threat_manager):
         """Test successful comprehensive IP enrichment."""
         # Use a unique IP to avoid cache hits
         unique_ip = f"10.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}"
-        
+
         # Mock DShield client response
         mock_response = {
             "reputation_score": 85.0,
@@ -119,15 +143,15 @@ class TestThreatIntelligenceManager:
             "asn": "AS15169",
             "organization": "Google LLC",
             "attack_types": ["port_scan", "brute_force"],
-            "tags": ["malicious", "scanner"]
+            "tags": ["malicious", "scanner"],
         }
-        
+
         threat_manager.clients[ThreatIntelligenceSource.DSHIELD].get_ip_reputation = AsyncMock(
             return_value=mock_response
         )
-        
+
         result = await threat_manager.enrich_ip_comprehensive(unique_ip)
-        
+
         assert isinstance(result, ThreatIntelligenceResult)
         assert result.ip_address == unique_ip
         assert result.overall_threat_score == 15.0  # 100 - 85
@@ -137,58 +161,63 @@ class TestThreatIntelligenceManager:
         # assert len(result.threat_indicators) > 0
         assert result.geographic_data["country"] == "US"
         assert result.network_data["asn"] == "AS15169"
-    
+
     @pytest.mark.asyncio
     async def test_enrich_ip_comprehensive_invalid_ip(self, threat_manager):
         """Test IP enrichment with invalid IP address."""
         with pytest.raises(ValueError, match="Invalid IP address"):
             await threat_manager.enrich_ip_comprehensive("invalid_ip")
-    
+
     @pytest.mark.asyncio
     async def test_enrich_ip_comprehensive_no_sources(self):
         """Test IP enrichment when no sources are available."""
-        with patch('src.threat_intelligence_manager.get_config', return_value={
-            "threat_intelligence": {
-                "sources": {
-                    "dshield": {"enabled": False},
-                    "virustotal": {"enabled": False}
-                }
-            }
-        }), patch('src.threat_intelligence_manager.get_user_config', return_value=self.mock_user_config()):
-            
+        with (
+            patch(
+                'src.threat_intelligence_manager.get_config',
+                return_value={
+                    "threat_intelligence": {
+                        "sources": {"dshield": {"enabled": False}, "virustotal": {"enabled": False}}
+                    }
+                },
+            ),
+            patch(
+                'src.threat_intelligence_manager.get_user_config',
+                return_value=self.mock_user_config(),
+            ),
+        ):
             manager = ThreatIntelligenceManager()
-            
+
             with pytest.raises(RuntimeError, match="No threat intelligence sources available"):
                 await manager.enrich_ip_comprehensive("8.8.8.8")
-            
+
             await manager.cleanup()
-    
+
     @pytest.mark.asyncio
     async def test_enrich_ip_comprehensive_cache_hit(self, threat_manager):
         """Test IP enrichment with cache hit."""
         # Use a unique IP to avoid cache persistence from other tests
         unique_ip = f"192.168.{hash(str(threat_manager)) % 255}.1"
-        
+
         # First call to populate cache
         mock_response = {"reputation_score": 50.0}
         threat_manager.clients[ThreatIntelligenceSource.DSHIELD].get_ip_reputation = AsyncMock(
             return_value=mock_response
         )
-        
+
         result1 = await threat_manager.enrich_ip_comprehensive(unique_ip)
         assert not result1.cache_hit
-        
+
         # Second call should hit cache
         result2 = await threat_manager.enrich_ip_comprehensive(unique_ip)
         assert result2.cache_hit
         assert result2.overall_threat_score == result1.overall_threat_score
-    
+
     @pytest.mark.asyncio
     async def test_enrich_ip_comprehensive_none_reputation(self, threat_manager):
         """Test IP enrichment when reputation score is None."""
         # Use a unique IP to avoid cache persistence from other tests
         unique_ip = f"203.0.{hash(str(threat_manager)) % 255}.1"
-        
+
         # Mock DShield client response with None reputation score
         mock_response = {
             "reputation_score": None,
@@ -196,46 +225,46 @@ class TestThreatIntelligenceManager:
             "asn": "AS15169",
             "organization": "Google LLC",
             "attack_types": [],
-            "tags": []
+            "tags": [],
         }
-        
+
         threat_manager.clients[ThreatIntelligenceSource.DSHIELD].get_ip_reputation = AsyncMock(
             return_value=mock_response
         )
-        
+
         result = await threat_manager.enrich_ip_comprehensive(unique_ip)
-        
+
         assert isinstance(result, ThreatIntelligenceResult)
         assert result.ip_address == unique_ip
         assert result.overall_threat_score is None  # No valid reputation score
         assert result.confidence_score is None  # No valid reputation score
         assert ThreatIntelligenceSource.DSHIELD in result.sources_queried
-    
+
     @pytest.mark.asyncio
     async def test_enrich_domain_comprehensive(self, threat_manager):
         """Test comprehensive domain enrichment."""
         result = await threat_manager.enrich_domain_comprehensive("example.com")
-        
+
         assert isinstance(result, DomainIntelligence)
         assert result.domain == "example.com"
         assert result.sources_queried == []  # No sources implemented yet
-    
+
     @pytest.mark.asyncio
     async def test_enrich_domain_comprehensive_invalid_domain(self, threat_manager):
         """Test domain enrichment with invalid domain."""
         with pytest.raises(ValueError, match="Invalid domain"):
             await threat_manager.enrich_domain_comprehensive("")
-        
+
         with pytest.raises(ValueError, match="Invalid domain"):
             await threat_manager.enrich_domain_comprehensive("nodots")
-    
+
     @pytest.mark.asyncio
     async def test_correlate_threat_indicators(self, threat_manager):
         """Test threat indicator correlation."""
         indicators = ["8.8.8.8", "example.com", "malware.exe"]
-        
+
         result = await threat_manager.correlate_threat_indicators(indicators)
-        
+
         assert isinstance(result, dict)
         assert "correlation_id" in result
         assert "indicators" in result
@@ -243,13 +272,13 @@ class TestThreatIntelligenceManager:
         assert "correlations" in result
         assert "relationships" in result
         assert "confidence_score" in result
-    
+
     @pytest.mark.asyncio
     async def test_correlate_threat_indicators_empty_list(self, threat_manager):
         """Test threat indicator correlation with empty list."""
         with pytest.raises(ValueError, match="Indicators list cannot be empty"):
             await threat_manager.correlate_threat_indicators([])
-    
+
     def test_classify_indicator(self, threat_manager):
         """Test indicator classification."""
         assert threat_manager._classify_indicator("8.8.8.8") == "ip_address"
@@ -257,51 +286,55 @@ class TestThreatIntelligenceManager:
         assert threat_manager._classify_indicator("a" * 32) == "hash"
         assert threat_manager._classify_indicator("CVE-2021-1234") == "cve"
         assert threat_manager._classify_indicator("generic_indicator") == "generic"
-    
+
+    @pytest.mark.slow
     def test_deduplicate_indicators(self, threat_manager):
         """Test indicator deduplication."""
         indicators = ["malware", "MALWARE", "port_scan", "malware"]
-        
+
         result = threat_manager._deduplicate_indicators(indicators)
-        
+
         assert len(result) == 2  # malware and port_scan
         malware_indicators = [ind for ind in result if ind["indicator"] == "malware"]
         assert len(malware_indicators) == 1
-        assert malware_indicators[0]["count"] == 3  # "malware", "MALWARE", "malware" = 3 occurrences
-    
+        assert (
+            malware_indicators[0]["count"] == 3
+        )  # "malware", "MALWARE", "malware" = 3 occurrences
+
     @pytest.mark.asyncio
+    @pytest.mark.slow
     async def test_rate_limiting(self, threat_manager):
         """Test rate limiting functionality."""
         source = ThreatIntelligenceSource.DSHIELD
-        
+
         # Should not raise exception for first request
         await threat_manager._check_rate_limit(source)
-        
+
         # Mock many requests to exceed rate limit (use old timestamps to force waiting)
         old_timestamp = datetime.now().timestamp() - 30  # 30 seconds ago
         threat_manager.rate_limit_trackers[source] = [old_timestamp] * 60
-        
+
         # The enhanced rate limiting should wait instead of immediately raising
         # Let's test that it doesn't raise immediately
         start_time = datetime.now()
         await threat_manager._check_rate_limit(source)
         end_time = datetime.now()
-        
+
         # Should have waited some time (at least 1 second)
         wait_duration = (end_time - start_time).total_seconds()
         assert wait_duration > 0.5  # Should have waited at least 0.5 seconds
-    
+
     def test_get_available_sources(self, threat_manager):
         """Test getting available sources."""
         sources = threat_manager.get_available_sources()
-        
+
         assert isinstance(sources, list)
         assert ThreatIntelligenceSource.DSHIELD in sources
-    
+
     def test_get_source_status(self, threat_manager):
         """Test source status retrieval."""
         status = threat_manager.get_source_status()
-        
+
         assert "dshield" in status
         assert status["dshield"]["enabled"] is True
         assert status["dshield"]["client_type"] == "DShieldClient"
@@ -311,10 +344,10 @@ class TestThreatIntelligenceManager:
     async def test_cache_behavior(self, threat_manager):
         """Test that repeated enrichment uses cache."""
         ip = "8.8.8.8"
-        result1 = await threat_manager.enrich_ip_comprehensive(ip)
+        await threat_manager.enrich_ip_comprehensive(ip)
         result2 = await threat_manager.enrich_ip_comprehensive(ip)
         assert result2.cache_hit
-    
+
     @pytest.mark.asyncio
     async def test_sqlite_cache_initialization(self, threat_manager):
         """Test SQLite cache initialization."""
@@ -322,82 +355,82 @@ class TestThreatIntelligenceManager:
         assert hasattr(threat_manager, 'sqlite_cache_enabled')
         assert hasattr(threat_manager, 'sqlite_cache_path')
         assert hasattr(threat_manager, 'sqlite_cache_ttl')
-        
+
         # Check if database file exists (if enabled)
         if threat_manager.sqlite_cache_enabled:
             assert os.path.exists(threat_manager.sqlite_cache_path)
-    
+
     @pytest.mark.asyncio
     async def test_sqlite_cache_storage_and_retrieval(self, threat_manager):
         """Test SQLite cache storage and retrieval."""
         if not threat_manager.sqlite_cache_enabled:
             pytest.skip("SQLite cache not enabled")
-        
+
         # Use a unique IP to avoid cache persistence from other tests
         unique_ip = f"10.0.{hash(str(threat_manager)) % 255}.1"
-        
+
         # First call should populate cache
         result1 = await threat_manager.enrich_ip_comprehensive(unique_ip)
         assert not result1.cache_hit
-        
+
         # Second call should hit SQLite cache
         result2 = await threat_manager.enrich_ip_comprehensive(unique_ip)
         assert result2.cache_hit
-        
+
         # Verify results are the same
         assert result1.ip_address == result2.ip_address
         assert result1.overall_threat_score == result2.overall_threat_score
-    
+
     @pytest.mark.asyncio
     async def test_sqlite_cache_expiry(self, threat_manager):
         """Test SQLite cache expiry behavior."""
         if not threat_manager.sqlite_cache_enabled:
             pytest.skip("SQLite cache not enabled")
-        
+
         # Use a unique IP to avoid cache persistence from other tests
         unique_ip = f"172.16.{hash(str(threat_manager)) % 255}.1"
-        
+
         # Temporarily set a very short TTL for testing
         original_ttl = threat_manager.sqlite_cache_ttl
         threat_manager.sqlite_cache_ttl = timedelta(seconds=1)
-        
+
         try:
             # Clear memory cache to ensure we're testing SQLite cache
             threat_manager.cache.clear()
-            
+
             # First call to populate cache
             result1 = await threat_manager.enrich_ip_comprehensive(unique_ip)
             assert not result1.cache_hit
-            
+
             # Wait for cache to expire
             await asyncio.sleep(2)
-            
+
             # Clear memory cache again to force SQLite lookup
             threat_manager.cache.clear()
-            
+
             # Second call should not hit cache due to expiry
             result2 = await threat_manager.enrich_ip_comprehensive(unique_ip)
             assert not result2.cache_hit
         finally:
             # Restore original TTL
             threat_manager.sqlite_cache_ttl = original_ttl
-    
+
     def test_get_cache_statistics(self, threat_manager):
         """Test cache statistics retrieval."""
         stats = threat_manager.get_cache_statistics()
-        
+
         # Check memory cache stats
         assert "memory_cache" in stats
         assert stats["memory_cache"]["enabled"] is True
         assert "size" in stats["memory_cache"]
         assert "ttl_hours" in stats["memory_cache"]
-        
+
         # Check SQLite cache stats
         assert "sqlite_cache" in stats
         assert "enabled" in stats["sqlite_cache"]
         assert "path" in stats["sqlite_cache"]
         assert "ttl_hours" in stats["sqlite_cache"]
-        
+
         # If SQLite cache is enabled, check additional stats
         if stats["sqlite_cache"]["enabled"]:
             assert "total_entries" in stats["sqlite_cache"]
@@ -408,66 +441,60 @@ class TestThreatIntelligenceManager:
 
 class TestThreatIntelligenceModels:
     """Test suite for threat intelligence data models."""
-    
+
     def test_threat_intelligence_result_creation(self):
         """Test ThreatIntelligenceResult model creation."""
         result = ThreatIntelligenceResult(
-            ip_address="8.8.8.8",
-            overall_threat_score=25.0,
-            confidence_score=0.8
+            ip_address="8.8.8.8", overall_threat_score=25.0, confidence_score=0.8
         )
-        
+
         assert result.ip_address == "8.8.8.8"
         assert result.overall_threat_score == 25.0
         assert result.confidence_score == 0.8
         assert isinstance(result.query_timestamp, datetime)
         assert result.cache_hit is False
-    
+
     def test_threat_intelligence_result_validation(self):
         """Test ThreatIntelligenceResult validation."""
         # Valid IP
         result = ThreatIntelligenceResult(ip_address="8.8.8.8")
         assert result.ip_address == "8.8.8.8"
-        
+
         # Invalid IP
         with pytest.raises(ValueError, match="Invalid IP address"):
             ThreatIntelligenceResult(ip_address="invalid_ip")
-        
+
         # Invalid threat score
         with pytest.raises(ValueError, match="Overall threat score must be between 0 and 100"):
             ThreatIntelligenceResult(ip_address="8.8.8.8", overall_threat_score=150.0)
-        
+
         # Invalid confidence score
         with pytest.raises(ValueError, match="Confidence score must be between 0 and 1"):
             ThreatIntelligenceResult(ip_address="8.8.8.8", confidence_score=1.5)
-    
+
     def test_domain_intelligence_creation(self):
         """Test DomainIntelligence model creation."""
-        result = DomainIntelligence(
-            domain="example.com",
-            threat_score=30.0,
-            reputation_score=70.0
-        )
-        
+        result = DomainIntelligence(domain="example.com", threat_score=30.0, reputation_score=70.0)
+
         assert result.domain == "example.com"
         assert result.threat_score == 30.0
         assert result.reputation_score == 70.0
         assert isinstance(result.query_timestamp, datetime)
         assert result.cache_hit is False
-    
+
     def test_domain_intelligence_validation(self):
         """Test DomainIntelligence validation."""
         # Valid domain
         result = DomainIntelligence(domain="example.com")
         assert result.domain == "example.com"
-        
+
         # Invalid domain
         with pytest.raises(ValueError, match="Invalid domain"):
             DomainIntelligence(domain="")
-        
+
         with pytest.raises(ValueError, match="Invalid domain"):
             DomainIntelligence(domain="nodots")
-        
+
         # Invalid scores
         with pytest.raises(ValueError, match="Score must be between 0 and 100"):
             DomainIntelligence(domain="example.com", threat_score=150.0)
@@ -475,7 +502,7 @@ class TestThreatIntelligenceModels:
 
 class TestThreatIntelligenceSource:
     """Test suite for ThreatIntelligenceSource enum."""
-    
+
     def test_source_values(self):
         """Test threat intelligence source values."""
         assert ThreatIntelligenceSource.DSHIELD.value == "dshield"
@@ -484,7 +511,7 @@ class TestThreatIntelligenceSource:
         assert ThreatIntelligenceSource.ABUSEIPDB.value == "abuseipdb"
         assert ThreatIntelligenceSource.ALIENVAULT.value == "alienvault"
         assert ThreatIntelligenceSource.THREATFOX.value == "threatfox"
-    
+
     def test_source_enumeration(self):
         """Test threat intelligence source enumeration."""
         sources = list(ThreatIntelligenceSource)
@@ -495,30 +522,34 @@ class TestThreatIntelligenceSource:
 
 class TestThreatIntelligenceIntegration:
     """Integration tests for threat intelligence functionality."""
-    
+
     @pytest_asyncio.fixture
     async def manager(self) -> ThreatIntelligenceManager:
         """Create a Threat Intelligence Manager instance for integration testing."""
         manager = ThreatIntelligenceManager()
         yield manager
         await manager.cleanup()
-        
+
         # Clean up SQLite cache file after each test
-        if hasattr(manager, 'sqlite_cache_enabled') and manager.sqlite_cache_enabled and os.path.exists(manager.sqlite_cache_path):
+        if (
+            hasattr(manager, 'sqlite_cache_enabled')
+            and manager.sqlite_cache_enabled
+            and os.path.exists(manager.sqlite_cache_path)
+        ):
             try:
                 os.remove(manager.sqlite_cache_path)
             except Exception:
                 pass  # Ignore cleanup errors
-    
+
     @pytest.mark.asyncio
     async def test_manager_context_manager(self):
         """Test Threat Intelligence Manager as context manager."""
         async with ThreatIntelligenceManager() as manager:
             assert isinstance(manager, ThreatIntelligenceManager)
             assert manager.clients is not None
-        
+
         # Manager should be cleaned up after context exit
-    
+
     @pytest.mark.asyncio
     async def test_correlation_with_multiple_sources(self):
         """Test correlation with multiple sources (when implemented)."""
@@ -526,40 +557,44 @@ class TestThreatIntelligenceIntegration:
         async with ThreatIntelligenceManager() as manager:
             # For now, test with just DShield
             result = await manager.enrich_ip_comprehensive("8.8.8.8")
-            
+
             assert isinstance(result, ThreatIntelligenceResult)
             assert result.ip_address == "8.8.8.8"
             assert len(result.sources_queried) >= 0  # At least 0 sources
-    
+
     @pytest.mark.asyncio
     async def test_cache_eviction(self):
         """Test cache eviction behavior."""
+
         def mock_user_config():
             """Create a mock user config for this test."""
             config = MagicMock()
             config.get_setting.return_value = "default_value"
-            
+
             # Mock performance settings for SQLite cache
             performance_settings = MagicMock()
             performance_settings.enable_sqlite_cache = True
             performance_settings.sqlite_cache_ttl_hours = 24
             performance_settings.sqlite_cache_db_name = "test_enrichment_cache.sqlite3"
             config.performance_settings = performance_settings
-            
+
             # Mock database path methods
             config.get_database_directory.return_value = "/tmp/test_db"
-            config.get_cache_database_path.return_value = "/tmp/test_db/test_enrichment_cache.sqlite3"
-            
+            config.get_cache_database_path.return_value = (
+                "/tmp/test_db/test_enrichment_cache.sqlite3"
+            )
+
             return config
-        
-        with patch('src.threat_intelligence_manager.get_config', return_value={
-            "threat_intelligence": {
-                "sources": {
-                    "dshield": {"enabled": True}
-                }
-            }
-        }), patch('src.threat_intelligence_manager.get_user_config', return_value=mock_user_config()):
-            
+
+        with (
+            patch(
+                'src.threat_intelligence_manager.get_config',
+                return_value={"threat_intelligence": {"sources": {"dshield": {"enabled": True}}}},
+            ),
+            patch(
+                'src.threat_intelligence_manager.get_user_config', return_value=mock_user_config()
+            ),
+        ):
             async with ThreatIntelligenceManager() as manager:
                 # Test that cache eviction works properly
                 # This is a basic test to ensure the manager can be created and cleaned up
@@ -572,52 +607,55 @@ class TestThreatIntelligenceIntegration:
         # Mock config to enable writeback and DShield source
         mock_config = {
             "threat_intelligence": {
-                "sources": {
-                    "dshield": {"enabled": True, "rate_limit_requests_per_minute": 60}
-                },
+                "sources": {"dshield": {"enabled": True, "rate_limit_requests_per_minute": 60}},
                 "elasticsearch": {
                     "enabled": True,
                     "writeback_enabled": True,
                     "hosts": ["http://localhost:9200"],
-                    "index_prefix": "enrichment-intel"
-                }
+                    "index_prefix": "enrichment-intel",
+                },
             }
         }
-        
+
         # Mock user config
         mock_user_config = MagicMock()
         mock_user_config.performance_settings.enable_sqlite_cache = True
         mock_user_config.performance_settings.sqlite_cache_ttl_hours = 24
         mock_user_config.performance_settings.sqlite_cache_db_name = "test_enrichment_cache.sqlite3"
         mock_user_config.get_database_directory.return_value = "/tmp/test_db"
-        mock_user_config.get_cache_database_path.return_value = "/tmp/test_db/test_enrichment_cache.sqlite3"
-        
-        with patch('src.threat_intelligence_manager.get_config', return_value=mock_config), \
-             patch('src.threat_intelligence_manager.get_user_config', return_value=mock_user_config):
-            
+        mock_user_config.get_cache_database_path.return_value = (
+            "/tmp/test_db/test_enrichment_cache.sqlite3"
+        )
+
+        with (
+            patch('src.threat_intelligence_manager.get_config', return_value=mock_config),
+            patch('src.threat_intelligence_manager.get_user_config', return_value=mock_user_config),
+        ):
             manager = ThreatIntelligenceManager()
-            
+
             # Mock Elasticsearch client
             mock_es_client = AsyncMock()
             manager.elasticsearch_client = mock_es_client
-            
+
             # Mock DShield client
             manager.clients[ThreatIntelligenceSource.DSHIELD] = AsyncMock()
             manager.clients[ThreatIntelligenceSource.DSHIELD].get_ip_reputation.return_value = {
                 "threat_score": 75.0,
-                "confidence": 0.8
+                "confidence": 0.8,
             }
-            
+
             # Use a unique IP to avoid cache hits
-            unique_ip = f"10.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}"
-            
+            unique_ip = (
+                f"10.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}"
+            )
+
             # Perform enrichment
             result = await manager.enrich_ip_comprehensive(unique_ip)
-            
+
             # Verify Elasticsearch write was called
             mock_es_client.index.assert_called_once()
             call_args = mock_es_client.index.call_args
-            
+
             # Verify document structure
             doc = call_args[1]["document"]
             assert doc["indicator"] == unique_ip
@@ -625,68 +663,71 @@ class TestThreatIntelligenceIntegration:
             assert "sources" in doc
             assert "timestamp" in doc
             assert doc["threat_score"] == result.overall_threat_score
-            
+
             # Verify index naming
             index_name = call_args[1]["index"]
             assert index_name.startswith("enrichment-intel-")
-            
+
             await manager.cleanup()
-    
+
     @pytest.mark.asyncio
     async def test_elasticsearch_writeback_disabled(self):
         """Test Elasticsearch writeback when disabled."""
         # Mock config to disable writeback but enable DShield source
         mock_config = {
             "threat_intelligence": {
-                "sources": {
-                    "dshield": {"enabled": True, "rate_limit_requests_per_minute": 60}
-                },
+                "sources": {"dshield": {"enabled": True, "rate_limit_requests_per_minute": 60}},
                 "elasticsearch": {
                     "enabled": True,
                     "writeback_enabled": False,  # Disabled
                     "hosts": ["http://localhost:9200"],
-                    "index_prefix": "enrichment-intel"
-                }
+                    "index_prefix": "enrichment-intel",
+                },
             }
         }
-        
+
         # Mock user config
         mock_user_config = MagicMock()
         mock_user_config.performance_settings.enable_sqlite_cache = True
         mock_user_config.performance_settings.sqlite_cache_ttl_hours = 24
         mock_user_config.performance_settings.sqlite_cache_db_name = "test_enrichment_cache.sqlite3"
         mock_user_config.get_database_directory.return_value = "/tmp/test_db"
-        mock_user_config.get_cache_database_path.return_value = "/tmp/test_db/test_enrichment_cache.sqlite3"
-        
-        with patch('src.threat_intelligence_manager.get_config', return_value=mock_config), \
-             patch('src.threat_intelligence_manager.get_user_config', return_value=mock_user_config):
-            
+        mock_user_config.get_cache_database_path.return_value = (
+            "/tmp/test_db/test_enrichment_cache.sqlite3"
+        )
+
+        with (
+            patch('src.threat_intelligence_manager.get_config', return_value=mock_config),
+            patch('src.threat_intelligence_manager.get_user_config', return_value=mock_user_config),
+        ):
             manager = ThreatIntelligenceManager()
-            
+
             # Mock Elasticsearch client
             mock_es_client = AsyncMock()
             manager.elasticsearch_client = mock_es_client
-            
+
             # Mock DShield client
             manager.clients[ThreatIntelligenceSource.DSHIELD] = AsyncMock()
             manager.clients[ThreatIntelligenceSource.DSHIELD].get_ip_reputation.return_value = {
                 "threat_score": 75.0,
-                "confidence": 0.8
+                "confidence": 0.8,
             }
-            
+
             # Use a unique IP to avoid cache hits
-            unique_ip = f"10.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}"
-            
+            unique_ip = (
+                f"10.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}"
+            )
+
             # Perform enrichment
             result = await manager.enrich_ip_comprehensive(unique_ip)
-            
+
             # Verify Elasticsearch write was NOT called
             mock_es_client.index.assert_not_called()
-            
+
             # Verify enrichment still works
             assert result.ip_address == unique_ip
             assert len(result.sources_queried) == 1
-            
+
             await manager.cleanup()
 
     @pytest.mark.asyncio
@@ -695,51 +736,54 @@ class TestThreatIntelligenceIntegration:
         # Mock config to enable writeback and DShield source
         mock_config = {
             "threat_intelligence": {
-                "sources": {
-                    "dshield": {"enabled": True, "rate_limit_requests_per_minute": 60}
-                },
+                "sources": {"dshield": {"enabled": True, "rate_limit_requests_per_minute": 60}},
                 "elasticsearch": {
                     "enabled": True,
                     "writeback_enabled": True,
                     "hosts": ["http://localhost:9200"],
-                    "index_prefix": "enrichment-intel"
-                }
+                    "index_prefix": "enrichment-intel",
+                },
             }
         }
-        
+
         # Mock user config
         mock_user_config = MagicMock()
         mock_user_config.performance_settings.enable_sqlite_cache = True
         mock_user_config.performance_settings.sqlite_cache_ttl_hours = 24
         mock_user_config.performance_settings.sqlite_cache_db_name = "test_enrichment_cache.sqlite3"
         mock_user_config.get_database_directory.return_value = "/tmp/test_db"
-        mock_user_config.get_cache_database_path.return_value = "/tmp/test_db/test_enrichment_cache.sqlite3"
-        
-        with patch('src.threat_intelligence_manager.get_config', return_value=mock_config), \
-             patch('src.threat_intelligence_manager.get_user_config', return_value=mock_user_config):
-            
+        mock_user_config.get_cache_database_path.return_value = (
+            "/tmp/test_db/test_enrichment_cache.sqlite3"
+        )
+
+        with (
+            patch('src.threat_intelligence_manager.get_config', return_value=mock_config),
+            patch('src.threat_intelligence_manager.get_user_config', return_value=mock_user_config),
+        ):
             manager = ThreatIntelligenceManager()
-            
+
             # Ensure no Elasticsearch client
             manager.elasticsearch_client = None
-            
+
             # Mock DShield client
             manager.clients[ThreatIntelligenceSource.DSHIELD] = AsyncMock()
             manager.clients[ThreatIntelligenceSource.DSHIELD].get_ip_reputation.return_value = {
                 "threat_score": 75.0,
-                "confidence": 0.8
+                "confidence": 0.8,
             }
-            
+
             # Use a unique IP to avoid cache hits
-            unique_ip = f"10.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}"
-            
+            unique_ip = (
+                f"10.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}"
+            )
+
             # Perform enrichment - should not fail
             result = await manager.enrich_ip_comprehensive(unique_ip)
-            
+
             # Verify enrichment still works
             assert result.ip_address == unique_ip
             assert len(result.sources_queried) == 1
-            
+
             await manager.cleanup()
 
     @pytest.mark.asyncio
@@ -748,56 +792,59 @@ class TestThreatIntelligenceIntegration:
         # Mock config to enable writeback and DShield source
         mock_config = {
             "threat_intelligence": {
-                "sources": {
-                    "dshield": {"enabled": True, "rate_limit_requests_per_minute": 60}
-                },
+                "sources": {"dshield": {"enabled": True, "rate_limit_requests_per_minute": 60}},
                 "elasticsearch": {
                     "enabled": True,
                     "writeback_enabled": True,
                     "hosts": ["http://localhost:9200"],
-                    "index_prefix": "enrichment-intel"
-                }
+                    "index_prefix": "enrichment-intel",
+                },
             }
         }
-        
+
         # Mock user config
         mock_user_config = MagicMock()
         mock_user_config.performance_settings.enable_sqlite_cache = True
         mock_user_config.performance_settings.sqlite_cache_ttl_hours = 24
         mock_user_config.performance_settings.sqlite_cache_db_name = "test_enrichment_cache.sqlite3"
         mock_user_config.get_database_directory.return_value = "/tmp/test_db"
-        mock_user_config.get_cache_database_path.return_value = "/tmp/test_db/test_enrichment_cache.sqlite3"
-        
-        with patch('src.threat_intelligence_manager.get_config', return_value=mock_config), \
-             patch('src.threat_intelligence_manager.get_user_config', return_value=mock_user_config):
-            
+        mock_user_config.get_cache_database_path.return_value = (
+            "/tmp/test_db/test_enrichment_cache.sqlite3"
+        )
+
+        with (
+            patch('src.threat_intelligence_manager.get_config', return_value=mock_config),
+            patch('src.threat_intelligence_manager.get_user_config', return_value=mock_user_config),
+        ):
             manager = ThreatIntelligenceManager()
-            
+
             # Mock Elasticsearch client that raises an exception
             mock_es_client = AsyncMock()
             mock_es_client.index.side_effect = Exception("Elasticsearch connection failed")
             manager.elasticsearch_client = mock_es_client
-            
+
             # Mock DShield client
             manager.clients[ThreatIntelligenceSource.DSHIELD] = AsyncMock()
             manager.clients[ThreatIntelligenceSource.DSHIELD].get_ip_reputation.return_value = {
                 "threat_score": 75.0,
-                "confidence": 0.8
+                "confidence": 0.8,
             }
-            
+
             # Use a unique IP to avoid cache hits
-            unique_ip = f"10.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}"
-            
+            unique_ip = (
+                f"10.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}"
+            )
+
             # Perform enrichment - should not fail due to ES error
             result = await manager.enrich_ip_comprehensive(unique_ip)
-            
+
             # Verify enrichment still works despite ES error
             assert result.ip_address == unique_ip
             assert len(result.sources_queried) == 1
-            
+
             # Verify Elasticsearch write was attempted
             mock_es_client.index.assert_called_once()
-            
+
             await manager.cleanup()
 
     @pytest.mark.asyncio
@@ -806,54 +853,57 @@ class TestThreatIntelligenceIntegration:
         # Mock config to enable writeback and DShield source
         mock_config = {
             "threat_intelligence": {
-                "sources": {
-                    "dshield": {"enabled": True, "rate_limit_requests_per_minute": 60}
-                },
+                "sources": {"dshield": {"enabled": True, "rate_limit_requests_per_minute": 60}},
                 "elasticsearch": {
                     "enabled": True,
                     "writeback_enabled": True,
                     "hosts": ["http://localhost:9200"],
-                    "index_prefix": "enrichment-intel"
-                }
+                    "index_prefix": "enrichment-intel",
+                },
             }
         }
-        
+
         # Mock user config
         mock_user_config = MagicMock()
         mock_user_config.performance_settings.enable_sqlite_cache = True
         mock_user_config.performance_settings.sqlite_cache_ttl_hours = 24
         mock_user_config.performance_settings.sqlite_cache_db_name = "test_enrichment_cache.sqlite3"
         mock_user_config.get_database_directory.return_value = "/tmp/test_db"
-        mock_user_config.get_cache_database_path.return_value = "/tmp/test_db/test_enrichment_cache.sqlite3"
-        
-        with patch('src.threat_intelligence_manager.get_config', return_value=mock_config), \
-             patch('src.threat_intelligence_manager.get_user_config', return_value=mock_user_config):
-            
+        mock_user_config.get_cache_database_path.return_value = (
+            "/tmp/test_db/test_enrichment_cache.sqlite3"
+        )
+
+        with (
+            patch('src.threat_intelligence_manager.get_config', return_value=mock_config),
+            patch('src.threat_intelligence_manager.get_user_config', return_value=mock_user_config),
+        ):
             manager = ThreatIntelligenceManager()
-            
+
             # Mock Elasticsearch client
             mock_es_client = AsyncMock()
             manager.elasticsearch_client = mock_es_client
-            
+
             # Mock DShield client with rich data
             manager.clients[ThreatIntelligenceSource.DSHIELD] = AsyncMock()
             manager.clients[ThreatIntelligenceSource.DSHIELD].get_ip_reputation.return_value = {
                 "threat_score": 75.0,
                 "confidence": 0.8,
                 "asn": 15169,
-                "country": "US"
+                "country": "US",
             }
-            
+
             # Use a unique IP to avoid cache hits
-            unique_ip = f"10.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}"
-            
+            unique_ip = (
+                f"10.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}"
+            )
+
             # Perform enrichment
             result = await manager.enrich_ip_comprehensive(unique_ip)
-            
+
             # Get the document that was written to Elasticsearch
             call_args = mock_es_client.index.call_args
             doc = call_args[1]["document"]
-            
+
             # Verify required fields
             assert doc["indicator"] == unique_ip
             assert doc["indicator_type"] == "ip"
@@ -861,24 +911,24 @@ class TestThreatIntelligenceIntegration:
             assert "timestamp" in doc
             assert "threat_score" in doc
             assert "confidence_score" in doc
-            
+
             # Verify sources data
             sources = doc["sources"]
             assert ThreatIntelligenceSource.DSHIELD.value in sources
-            
+
             # Verify network data if present
             if result.network_data:
                 assert doc["asn"] == result.network_data.get("asn")
-            
+
             # Verify geographic data if present
             if result.geographic_data:
                 assert doc["geo"] == result.geographic_data
-            
+
             # Verify tags from threat indicators
             if result.threat_indicators:
                 assert "tags" in doc
                 assert isinstance(doc["tags"], list)
-            
+
             await manager.cleanup()
 
     @pytest.mark.asyncio
@@ -887,56 +937,59 @@ class TestThreatIntelligenceIntegration:
         # Mock config with custom index prefix
         mock_config = {
             "threat_intelligence": {
-                "sources": {
-                    "dshield": {"enabled": True, "rate_limit_requests_per_minute": 60}
-                },
+                "sources": {"dshield": {"enabled": True, "rate_limit_requests_per_minute": 60}},
                 "elasticsearch": {
                     "enabled": True,
                     "writeback_enabled": True,
                     "hosts": ["http://localhost:9200"],
-                    "index_prefix": "custom-enrichment"
-                }
+                    "index_prefix": "custom-enrichment",
+                },
             }
         }
-        
+
         # Mock user config
         mock_user_config = MagicMock()
         mock_user_config.performance_settings.enable_sqlite_cache = True
         mock_user_config.performance_settings.sqlite_cache_ttl_hours = 24
         mock_user_config.performance_settings.sqlite_cache_db_name = "test_enrichment_cache.sqlite3"
         mock_user_config.get_database_directory.return_value = "/tmp/test_db"
-        mock_user_config.get_cache_database_path.return_value = "/tmp/test_db/test_enrichment_cache.sqlite3"
-        
-        with patch('src.threat_intelligence_manager.get_config', return_value=mock_config), \
-             patch('src.threat_intelligence_manager.get_user_config', return_value=mock_user_config):
-            
+        mock_user_config.get_cache_database_path.return_value = (
+            "/tmp/test_db/test_enrichment_cache.sqlite3"
+        )
+
+        with (
+            patch('src.threat_intelligence_manager.get_config', return_value=mock_config),
+            patch('src.threat_intelligence_manager.get_user_config', return_value=mock_user_config),
+        ):
             manager = ThreatIntelligenceManager()
-            
+
             # Mock Elasticsearch client
             mock_es_client = AsyncMock()
             manager.elasticsearch_client = mock_es_client
-            
+
             # Mock DShield client
             manager.clients[ThreatIntelligenceSource.DSHIELD] = AsyncMock()
             manager.clients[ThreatIntelligenceSource.DSHIELD].get_ip_reputation.return_value = {
                 "threat_score": 75.0,
-                "confidence": 0.8
+                "confidence": 0.8,
             }
-            
+
             # Use a unique IP to avoid cache hits
-            unique_ip = f"10.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}"
-            
+            unique_ip = (
+                f"10.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}"
+            )
+
             # Perform enrichment
             await manager.enrich_ip_comprehensive(unique_ip)
-            
+
             # Verify index naming
             call_args = mock_es_client.index.call_args
             index_name = call_args[1]["index"]
-            
+
             # Should follow pattern: prefix-YYYY.MM
             assert index_name.startswith("custom-enrichment-")
             assert len(index_name) == len("custom-enrichment-") + 7  # YYYY.MM format
-            
+
             await manager.cleanup()
 
     @pytest.mark.asyncio
@@ -945,56 +998,59 @@ class TestThreatIntelligenceIntegration:
         # Mock config to enable writeback and DShield source
         mock_config = {
             "threat_intelligence": {
-                "sources": {
-                    "dshield": {"enabled": True, "rate_limit_requests_per_minute": 60}
-                },
+                "sources": {"dshield": {"enabled": True, "rate_limit_requests_per_minute": 60}},
                 "elasticsearch": {
                     "enabled": True,
                     "writeback_enabled": True,
                     "hosts": ["http://localhost:9200"],
-                    "index_prefix": "enrichment-intel"
-                }
+                    "index_prefix": "enrichment-intel",
+                },
             }
         }
-        
+
         # Mock user config
         mock_user_config = MagicMock()
         mock_user_config.performance_settings.enable_sqlite_cache = True
         mock_user_config.performance_settings.sqlite_cache_ttl_hours = 24
         mock_user_config.performance_settings.sqlite_cache_db_name = "test_enrichment_cache.sqlite3"
         mock_user_config.get_database_directory.return_value = "/tmp/test_db"
-        mock_user_config.get_cache_database_path.return_value = "/tmp/test_db/test_enrichment_cache.sqlite3"
-        
-        with patch('src.threat_intelligence_manager.get_config', return_value=mock_config), \
-             patch('src.threat_intelligence_manager.get_user_config', return_value=mock_user_config):
-            
+        mock_user_config.get_cache_database_path.return_value = (
+            "/tmp/test_db/test_enrichment_cache.sqlite3"
+        )
+
+        with (
+            patch('src.threat_intelligence_manager.get_config', return_value=mock_config),
+            patch('src.threat_intelligence_manager.get_user_config', return_value=mock_user_config),
+        ):
             manager = ThreatIntelligenceManager()
-            
+
             # Mock Elasticsearch client
             mock_es_client = AsyncMock()
             manager.elasticsearch_client = mock_es_client
-            
+
             # Mock DShield client
             manager.clients[ThreatIntelligenceSource.DSHIELD] = AsyncMock()
             manager.clients[ThreatIntelligenceSource.DSHIELD].get_ip_reputation.return_value = {
                 "threat_score": 75.0,
-                "confidence": 0.8
+                "confidence": 0.8,
             }
-            
+
             # Use a unique IP to avoid cache hits
-            unique_ip = f"10.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}"
-            
+            unique_ip = (
+                f"10.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}"
+            )
+
             # Perform enrichment
             await manager.enrich_ip_comprehensive(unique_ip)
-            
+
             # Verify document ID format
             call_args = mock_es_client.index.call_args
             doc_id = call_args[1]["id"]
-            
+
             # Should be: ip_timestamp
             assert doc_id.startswith(f"{unique_ip}_")
             assert len(doc_id) > len(f"{unique_ip}_")
-            
+
             await manager.cleanup()
 
     @pytest.mark.asyncio
@@ -1003,66 +1059,67 @@ class TestThreatIntelligenceIntegration:
         # Mock config to enable writeback and DShield source
         mock_config = {
             "threat_intelligence": {
-                "sources": {
-                    "dshield": {"enabled": True, "rate_limit_requests_per_minute": 60}
-                },
+                "sources": {"dshield": {"enabled": True, "rate_limit_requests_per_minute": 60}},
                 "elasticsearch": {
                     "enabled": True,
                     "writeback_enabled": True,
                     "hosts": ["http://localhost:9200"],
-                    "index_prefix": "enrichment-intel"
-                }
+                    "index_prefix": "enrichment-intel",
+                },
             }
         }
-        
+
         # Mock user config
         mock_user_config = MagicMock()
         mock_user_config.performance_settings.enable_sqlite_cache = True
         mock_user_config.performance_settings.sqlite_cache_ttl_hours = 24
         mock_user_config.performance_settings.sqlite_cache_db_name = "test_enrichment_cache.sqlite3"
         mock_user_config.get_database_directory.return_value = "/tmp/test_db"
-        mock_user_config.get_cache_database_path.return_value = "/tmp/test_db/test_enrichment_cache.sqlite3"
-        
-        with patch('src.threat_intelligence_manager.get_config', return_value=mock_config), \
-             patch('src.threat_intelligence_manager.get_user_config', return_value=mock_user_config):
-            
+        mock_user_config.get_cache_database_path.return_value = (
+            "/tmp/test_db/test_enrichment_cache.sqlite3"
+        )
+
+        with (
+            patch('src.threat_intelligence_manager.get_config', return_value=mock_config),
+            patch('src.threat_intelligence_manager.get_user_config', return_value=mock_user_config),
+        ):
             manager = ThreatIntelligenceManager()
-            
+
             # Mock Elasticsearch client
             mock_es_client = AsyncMock()
             manager.elasticsearch_client = mock_es_client
-            
+
             # Mock DShield client
             manager.clients[ThreatIntelligenceSource.DSHIELD] = AsyncMock()
             manager.clients[ThreatIntelligenceSource.DSHIELD].get_ip_reputation.return_value = {
                 "threat_score": 75.0,
-                "confidence": 0.8
+                "confidence": 0.8,
             }
-            
+
             # Use unique IPs to avoid cache hits
             unique_ips = [
                 f"10.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}.{uuid.uuid4().int % 255}"
                 for _ in range(3)
             ]
-            
+
             # Perform multiple enrichments
             for ip in unique_ips:
                 await manager.enrich_ip_comprehensive(ip)
-            
+
             # Verify three documents were written
             assert mock_es_client.index.call_count == 3
-            
+
             # Verify different IPs
             calls = mock_es_client.index.call_args_list
             indicators = [call[1]["document"]["indicator"] for call in calls]
             assert unique_ips[0] in indicators
             assert unique_ips[1] in indicators
             assert unique_ips[2] in indicators
-            
+
             # Verify unique document IDs
             doc_ids = [call[1]["id"] for call in calls]
             assert len(set(doc_ids)) == 3  # All IDs should be unique
-            
+
             await manager.cleanup()
 
     @pytest.mark.asyncio
@@ -1077,7 +1134,7 @@ class TestThreatIntelligenceIntegration:
                 "attack_types": ["port_scan", "brute_force"],
                 "tags": ["malicious", "scanner"],
                 "country": "US",
-                "asn": "AS15169"
+                "asn": "AS15169",
             },
             ThreatIntelligenceSource.VIRUSTOTAL: {
                 "threat_score": 80.0,
@@ -1085,36 +1142,39 @@ class TestThreatIntelligenceIntegration:
                 "attack_types": ["port_scan"],
                 "tags": ["malicious"],
                 "country": "US",
-                "asn": "AS15169"
-            }
+                "asn": "AS15169",
+            },
         }
-        
+
         # Run enhanced correlation
         await manager._correlate_results(result)
-        
+
         # Verify weighted threat score (should be weighted by source reliability)
         # DShield: 75.0 * 0.8 = 60.0, VirusTotal: 80.0 * 0.9 = 72.0
         # Weighted average: (60.0 + 72.0) / (0.8 + 0.9) = 132.0 / 1.7 = 77.65
         assert result.overall_threat_score is not None
         assert 77.0 <= result.overall_threat_score <= 78.0
-        
+
         # Verify confidence score
         assert result.confidence_score is not None
         assert 0.8 <= result.confidence_score <= 0.9
-        
+
         # Verify correlated indicators
         assert len(result.threat_indicators) > 0
-        
+
         # Check that indicators are sorted by confidence
         if len(result.threat_indicators) > 1:
-            assert result.threat_indicators[0]['confidence'] >= result.threat_indicators[1]['confidence']
-        
+            assert (
+                result.threat_indicators[0]['confidence']
+                >= result.threat_indicators[1]['confidence']
+            )
+
         # Verify geographic data aggregation
         assert result.geographic_data.get('country') == 'US'
-        
+
         # Verify network data aggregation
         assert result.network_data.get('asn') == 'AS15169'
-        
+
         # Verify correlation metrics
         assert result.correlation_metrics is not None
         assert result.correlation_metrics['source_count'] == 2
@@ -1128,7 +1188,10 @@ class TestThreatIntelligenceIntegrationLive:
 
     @pytest.mark.integration
     @pytest.mark.asyncio
-    @pytest.mark.skipif(not os.getenv("RUN_LIVE_INTEGRATION_TESTS"), reason="Set RUN_LIVE_INTEGRATION_TESTS=1 to run live integration tests.")
+    @pytest.mark.skipif(
+        not os.getenv("RUN_LIVE_INTEGRATION_TESTS"),
+        reason="Set RUN_LIVE_INTEGRATION_TESTS=1 to run live integration tests.",
+    )
     async def test_real_api_enrichment(self):
         """Test real end-to-end enrichment with live DShield (and optionally other) APIs."""
         manager = ThreatIntelligenceManager()
@@ -1140,30 +1203,34 @@ class TestThreatIntelligenceIntegrationLive:
         assert result.overall_threat_score is not None or result.confidence_score is not None
         await manager.cleanup()
 
+
+@pytest.mark.slow
 class TestThreatIntelligenceIntegrationMocked:
     """Mocked end-to-end integration tests for CI/CD and isolated validation."""
 
     @pytest.mark.asyncio
     async def test_end_to_end_enrichment_and_writeback(self):
         """Test full enrichment, correlation, and writeback with all sources and ES mocked."""
-        with patch('src.threat_intelligence_manager.get_config') as mock_get_config, \
-             patch('src.threat_intelligence_manager.get_user_config') as mock_get_user_config:
+        with (
+            patch('src.threat_intelligence_manager.get_config') as mock_get_config,
+            patch('src.threat_intelligence_manager.get_user_config') as mock_get_user_config,
+        ):
             # Mock config with all sources enabled
             mock_get_config.return_value = {
                 "threat_intelligence": {
                     "sources": {
                         "dshield": {"enabled": True, "rate_limit_requests_per_minute": 60},
                         "virustotal": {"enabled": True, "rate_limit_requests_per_minute": 4},
-                        "shodan": {"enabled": True, "rate_limit_requests_per_minute": 60}
+                        "shodan": {"enabled": True, "rate_limit_requests_per_minute": 60},
                     },
                     "elasticsearch": {
                         "enabled": True,
                         "writeback_enabled": True,
                         "hosts": ["http://localhost:9200"],
-                        "index_prefix": "test-enrichment"
+                        "index_prefix": "test-enrichment",
                     },
                     "correlation": {"confidence_threshold": 0.7, "max_sources_per_query": 3},
-                    "cache_ttl_hours": 1
+                    "cache_ttl_hours": 1,
                 }
             }
             # Mock user config
@@ -1171,18 +1238,24 @@ class TestThreatIntelligenceIntegrationMocked:
             mock_uc.performance_settings.enable_sqlite_cache = True
             mock_uc.performance_settings.sqlite_cache_ttl_hours = 24
             mock_uc.get_database_directory.return_value = "/tmp/test_db"
-            mock_uc.get_cache_database_path.return_value = "/tmp/test_db/test_enrichment_cache.sqlite3"
+            mock_uc.get_cache_database_path.return_value = (
+                "/tmp/test_db/test_enrichment_cache.sqlite3"
+            )
             mock_get_user_config.return_value = mock_uc
 
             manager = ThreatIntelligenceManager()
             # Mock all source clients
-            for source in [ThreatIntelligenceSource.DSHIELD, ThreatIntelligenceSource.VIRUSTOTAL, ThreatIntelligenceSource.SHODAN]:
+            for source in [
+                ThreatIntelligenceSource.DSHIELD,
+                ThreatIntelligenceSource.VIRUSTOTAL,
+                ThreatIntelligenceSource.SHODAN,
+            ]:
                 mock_client = AsyncMock()
                 mock_client.get_ip_reputation.return_value = {
                     "threat_score": 80.0,
                     "confidence": 0.9,
                     "attack_types": ["port_scan"],
-                    "tags": ["malicious"]
+                    "tags": ["malicious"],
                 }
                 manager.clients[source] = mock_client
             # Mock Elasticsearch client
@@ -1205,23 +1278,24 @@ class TestThreatIntelligenceIntegrationMocked:
 
 class TestDiagnoseDataAvailability:
     """Test suite for the diagnose_data_availability method."""
-    
+
     @pytest.mark.asyncio
     async def test_diagnose_data_availability_success(self):
         """Test successful data availability diagnosis."""
-        with patch('src.threat_intelligence_manager.get_config') as mock_get_config, \
-             patch('src.threat_intelligence_manager.get_user_config') as mock_get_user_config, \
-             patch('src.elasticsearch_client.ElasticsearchClient') as mock_es_class:
-            
+        with (
+            patch('src.threat_intelligence_manager.get_config') as mock_get_config,
+            patch('src.threat_intelligence_manager.get_user_config') as mock_get_user_config,
+            patch('src.elasticsearch_client.ElasticsearchClient') as mock_es_class,
+        ):
             # Mock config
             mock_get_config.return_value = {
                 "threat_intelligence": {
                     "sources": {"dshield": {"enabled": True}},
                     "correlation": {"confidence_threshold": 0.7, "max_sources_per_query": 3},
-                    "cache_ttl_hours": 1
+                    "cache_ttl_hours": 1,
                 }
             }
-            
+
             # Mock user config with proper structure
             mock_uc = MagicMock()
             mock_uc.performance_settings.enable_sqlite_cache = False
@@ -1230,13 +1304,16 @@ class TestDiagnoseDataAvailability:
             mock_uc.get_database_directory.return_value = "/tmp/test_db"
             mock_uc.get_cache_database_path.return_value = "/tmp/test_db/test_cache.sqlite3"
             mock_get_user_config.return_value = mock_uc
-            
+
             # Mock Elasticsearch client
             mock_es_client = AsyncMock()
-            mock_es_client.get_available_indices.return_value = ["dshield-2024.01.01", "cowrie-2024.01.01"]
+            mock_es_client.get_available_indices.return_value = [
+                "dshield-2024.01.01",
+                "cowrie-2024.01.01",
+            ]
             mock_es_client.dshield_indices = ["dshield-*"]
             mock_es_client.fallback_indices = ["cowrie-*"]
-            
+
             # Mock index mapping
             mock_mapping = {
                 "dshield-2024.01.01": {
@@ -1244,59 +1321,60 @@ class TestDiagnoseDataAvailability:
                         "properties": {
                             "@timestamp": {"type": "date"},
                             "source.ip": {"type": "ip"},
-                            "destination.ip": {"type": "ip"}
+                            "destination.ip": {"type": "ip"},
                         }
                     }
                 }
             }
             mock_es_client.client.indices.get_mapping.return_value = mock_mapping
-            
+
             # Mock query results
             mock_es_client.query_dshield_events.return_value = ([{"id": "1"}], 100, {})
-            
+
             mock_es_class.return_value = mock_es_client
-            
+
             manager = ThreatIntelligenceManager()
             result = await manager.diagnose_data_availability()
-            
+
             # Verify the diagnosis structure
             assert isinstance(result, dict)
             assert "timestamp" in result
             assert "summary" in result
             assert "details" in result
             assert "recommendations" in result
-            
+
             # Verify summary
             assert result["summary"]["overall_status"] == "healthy"
             assert result["summary"]["severity"] == "low"
-            
+
             # Verify details
             assert "available_indices" in result["details"]
             assert "sample_mapping" in result["details"]
             assert "data_availability" in result["details"]
             assert "pattern_tests" in result["details"]
-            
+
             # Verify recommendations
             assert len(result["recommendations"]) > 0
-            
+
             await manager.cleanup()
-    
+
     @pytest.mark.asyncio
     async def test_diagnose_data_availability_no_indices(self):
         """Test data availability diagnosis when no indices are found."""
-        with patch('src.threat_intelligence_manager.get_config') as mock_get_config, \
-             patch('src.threat_intelligence_manager.get_user_config') as mock_get_user_config, \
-             patch('src.elasticsearch_client.ElasticsearchClient') as mock_es_class:
-            
+        with (
+            patch('src.threat_intelligence_manager.get_config') as mock_get_config,
+            patch('src.threat_intelligence_manager.get_user_config') as mock_get_user_config,
+            patch('src.elasticsearch_client.ElasticsearchClient') as mock_es_class,
+        ):
             # Mock config
             mock_get_config.return_value = {
                 "threat_intelligence": {
                     "sources": {"dshield": {"enabled": True}},
                     "correlation": {"confidence_threshold": 0.7, "max_sources_per_query": 3},
-                    "cache_ttl_hours": 1
+                    "cache_ttl_hours": 1,
                 }
             }
-            
+
             # Mock user config with proper structure
             mock_uc = MagicMock()
             mock_uc.performance_settings.enable_sqlite_cache = False
@@ -1305,46 +1383,47 @@ class TestDiagnoseDataAvailability:
             mock_uc.get_database_directory.return_value = "/tmp/test_db"
             mock_uc.get_cache_database_path.return_value = "/tmp/test_db/test_cache.sqlite3"
             mock_get_user_config.return_value = mock_uc
-            
+
             # Mock Elasticsearch client with no indices
             mock_es_client = AsyncMock()
             mock_es_client.get_available_indices.return_value = []
             mock_es_client.dshield_indices = ["dshield-*"]
             mock_es_client.fallback_indices = ["cowrie-*"]
-            
+
             mock_es_class.return_value = mock_es_client
-            
+
             manager = ThreatIntelligenceManager()
             result = await manager.diagnose_data_availability()
-            
+
             # Verify the diagnosis structure
             assert isinstance(result, dict)
             assert "issues_detected" in result["summary"]["overall_status"]
             assert result["summary"]["severity"] == "high"
-            
+
             # Verify recommendations for no indices
             recommendations = result["recommendations"]
             assert any("index_patterns" in rec for rec in recommendations)
             assert any("indices exist" in rec for rec in recommendations)
-            
+
             await manager.cleanup()
-    
+
     @pytest.mark.asyncio
     async def test_diagnose_data_availability_connection_error(self):
         """Test data availability diagnosis when Elasticsearch connection fails."""
-        with patch('src.threat_intelligence_manager.get_config') as mock_get_config, \
-             patch('src.threat_intelligence_manager.get_user_config') as mock_get_user_config, \
-             patch('src.elasticsearch_client.ElasticsearchClient') as mock_es_class:
-            
+        with (
+            patch('src.threat_intelligence_manager.get_config') as mock_get_config,
+            patch('src.threat_intelligence_manager.get_user_config') as mock_get_user_config,
+            patch('src.elasticsearch_client.ElasticsearchClient') as mock_es_class,
+        ):
             # Mock config
             mock_get_config.return_value = {
                 "threat_intelligence": {
                     "sources": {"dshield": {"enabled": True}},
                     "correlation": {"confidence_threshold": 0.7, "max_sources_per_query": 3},
-                    "cache_ttl_hours": 1
+                    "cache_ttl_hours": 1,
                 }
             }
-            
+
             # Mock user config with proper structure
             mock_uc = MagicMock()
             mock_uc.performance_settings.enable_sqlite_cache = False
@@ -1353,24 +1432,24 @@ class TestDiagnoseDataAvailability:
             mock_uc.get_database_directory.return_value = "/tmp/test_db"
             mock_uc.get_cache_database_path.return_value = "/tmp/test_db/test_cache.sqlite3"
             mock_get_user_config.return_value = mock_uc
-            
+
             # Mock Elasticsearch client that raises an error on connect
             mock_es_class.side_effect = Exception("Connection failed")
-            
+
             manager = ThreatIntelligenceManager()
             result = await manager.diagnose_data_availability()
-            
+
             # Verify the diagnosis structure
             assert isinstance(result, dict)
             assert result["summary"]["overall_status"] == "diagnosis_failed"
             assert result["summary"]["severity"] == "critical"
-            
+
             # Verify error details
             assert "diagnosis_error" in result["details"]
             assert "Connection failed" in result["details"]["diagnosis_error"]
-            
+
             await manager.cleanup()
 
 
 if __name__ == "__main__":
-    pytest.main([__file__, "-v"]) 
+    pytest.main([__file__, "-v"])
