@@ -8,22 +8,30 @@ error handling.
 
 import json
 import os
-import tempfile
 from pathlib import Path
-from typing import TYPE_CHECKING, Dict, Any
-from unittest.mock import AsyncMock, MagicMock, patch
+from typing import TYPE_CHECKING, Any
+from unittest.mock import MagicMock, Mock, patch
 
 import pytest
 import pytest_asyncio
 
 from src.latex_template_tools import LaTeXTemplateTools
 
+
+# Fixture to mock get_user_config for all tests
+@pytest.fixture(autouse=True)
+def mock_user_config():
+    """Mock get_user_config to prevent subprocess calls."""
+    with patch('src.latex_template_tools.get_user_config') as mock_get_user_config:
+        mock_user_config = Mock()
+        mock_user_config.output_directory = "/tmp/test_output"
+        mock_user_config.template_directory = "/tmp/test_templates"
+        mock_get_user_config.return_value = mock_user_config
+        yield mock_user_config
+
+
 if TYPE_CHECKING:
-    from _pytest.capture import CaptureFixture
-    from _pytest.fixtures import FixtureRequest
-    from _pytest.logging import LogCaptureFixture
-    from _pytest.monkeypatch import MonkeyPatch
-    from pytest_mock.plugin import MockerFixture
+    pass
 
 
 class TestLaTeXTemplateTools:
@@ -34,33 +42,29 @@ class TestLaTeXTemplateTools:
         """Create LaTeXTemplateTools instance with temporary template directory."""
         template_dir = tmp_path / "templates" / "Attack_Report"
         template_dir.mkdir(parents=True)
-        
+
         # Create template_info.json
         template_config = {
             "template_name": "DShield-MCP Attack Observation Report",
             "version": "2.1",
             "description": "Modular LaTeX template for DShield-MCP attack analysis reports",
-            "sections": [
-                "title_page",
-                "executive_summary",
-                "campaign_overview"
-            ],
+            "sections": ["title_page", "executive_summary", "campaign_overview"],
             "required_variables": {
                 "report_metadata": ["REPORT_NUMBER", "REPORT_TITLE", "AUTHOR_NAME", "REPORT_DATE"],
-                "analysis_data": ["CAMPAIGN_NAME", "TOTAL_SESSIONS", "MCP_VERSION"]
-            }
+                "analysis_data": ["CAMPAIGN_NAME", "TOTAL_SESSIONS", "MCP_VERSION"],
+            },
         }
-        
+
         with open(template_dir / "template_info.json", 'w') as f:
             json.dump(template_config, f)
-        
+
         # Create sample LaTeX files
         (template_dir / "main_report.tex").write_text(r"""
 % DShield-MCP Main Report Compilation File
 \input{preamble}
 \input{document_body}
 """)
-        
+
         (template_dir / "document_body.tex").write_text(r"""
 \begin{document}
 \input{sections/title_page}
@@ -68,43 +72,43 @@ class TestLaTeXTemplateTools:
 \input{sections/campaign_overview}
 \end{document}
 """)
-        
+
         (template_dir / "preamble.tex").write_text(r"""
 \documentclass{article}
 \title{{{REPORT_TITLE}}}
 \author{{{AUTHOR_NAME}}}
 \date{{{REPORT_DATE}}}
 """)
-        
+
         # Create sections directory
         sections_dir = template_dir / "sections"
         sections_dir.mkdir()
-        
+
         (sections_dir / "title_page.tex").write_text(r"""
 \maketitle
 Report Number: {REPORT_NUMBER}
 """)
-        
+
         (sections_dir / "executive_summary.tex").write_text(r"""
 \section{Executive Summary}
 Campaign: {CAMPAIGN_NAME}
 Total Sessions: {TOTAL_SESSIONS}
 MCP Version: {MCP_VERSION}
 """)
-        
+
         (sections_dir / "campaign_overview.tex").write_text(r"""
 \section{Campaign Overview}
 Analysis of campaign {CAMPAIGN_NAME}.
 """)
-        
+
         # Create assets directory to avoid errors
         assets_dir = template_dir / "assets"
         assets_dir.mkdir(exist_ok=True)
-        
+
         return LaTeXTemplateTools(str(template_dir))
 
     @pytest_asyncio.fixture
-    async def sample_document_data(self) -> Dict[str, Any]:
+    async def sample_document_data(self) -> dict[str, Any]:
         """Sample document data for testing."""
         return {
             "REPORT_NUMBER": "REP-2024-001",
@@ -113,7 +117,7 @@ Analysis of campaign {CAMPAIGN_NAME}.
             "REPORT_DATE": "2024-01-15",
             "CAMPAIGN_NAME": "Test Campaign",
             "TOTAL_SESSIONS": "150",
-            "MCP_VERSION": "1.0.0"
+            "MCP_VERSION": "1.0.0",
         }
 
     @pytest.mark.asyncio
@@ -121,7 +125,7 @@ Analysis of campaign {CAMPAIGN_NAME}.
         """Test initialization with valid template path."""
         template_dir = tmp_path / "templates" / "TestTemplate"
         template_dir.mkdir(parents=True)
-        
+
         tools = LaTeXTemplateTools(str(template_dir))
         assert tools.template_base_path == template_dir
 
@@ -129,7 +133,7 @@ Analysis of campaign {CAMPAIGN_NAME}.
     async def test_init_with_invalid_template_path(self, tmp_path: Path) -> None:
         """Test initialization with invalid template path."""
         invalid_path = tmp_path / "nonexistent"
-        
+
         with pytest.raises(FileNotFoundError, match="Template directory not found"):
             LaTeXTemplateTools(str(invalid_path))
 
@@ -137,11 +141,11 @@ Analysis of campaign {CAMPAIGN_NAME}.
     async def test_list_available_templates(self, latex_tools: LaTeXTemplateTools) -> None:
         """Test listing available templates."""
         result = await latex_tools.list_available_templates()
-        
+
         assert result["success"] is True
         assert len(result["templates"]) >= 1
         assert result["total_templates"] >= 1
-        
+
         # Check template metadata
         template = result["templates"][0]
         assert "name" in template
@@ -155,10 +159,10 @@ Analysis of campaign {CAMPAIGN_NAME}.
     async def test_get_template_schema(self, latex_tools: LaTeXTemplateTools) -> None:
         """Test getting template schema."""
         result = await latex_tools.get_template_schema("Attack_Report")
-        
+
         assert result["success"] is True
         assert result["schema"] is not None
-        
+
         schema = result["schema"]
         assert "template_name" in schema
         assert "version" in schema
@@ -170,140 +174,168 @@ Analysis of campaign {CAMPAIGN_NAME}.
         assert "example_data" in schema
 
     @pytest.mark.asyncio
-    async def test_get_template_schema_invalid_template(self, latex_tools: LaTeXTemplateTools) -> None:
+    async def test_get_template_schema_invalid_template(
+        self, latex_tools: LaTeXTemplateTools
+    ) -> None:
         """Test getting schema for invalid template."""
         result = await latex_tools.get_template_schema("InvalidTemplate")
-        
+
         assert result["success"] is False
         assert "error" in result
         assert "Template not found" in result["error"]
 
     @pytest.mark.asyncio
-    async def test_validate_document_data_valid(self, latex_tools: LaTeXTemplateTools, sample_document_data: Dict[str, Any]) -> None:
+    async def test_validate_document_data_valid(
+        self, latex_tools: LaTeXTemplateTools, sample_document_data: dict[str, Any]
+    ) -> None:
         """Test validation of valid document data."""
         result = await latex_tools.validate_document_data("Attack_Report", sample_document_data)
-        
+
         assert result["success"] is True
         assert result["validation"]["valid"] is True
         assert len(result["validation"]["errors"]) == 0
 
     @pytest.mark.asyncio
-    async def test_validate_document_data_missing_variables(self, latex_tools: LaTeXTemplateTools) -> None:
+    async def test_validate_document_data_missing_variables(
+        self, latex_tools: LaTeXTemplateTools
+    ) -> None:
         """Test validation of document data with missing variables."""
         incomplete_data = {
             "REPORT_TITLE": "Test Report",
-            "AUTHOR_NAME": "Test Analyst"
+            "AUTHOR_NAME": "Test Analyst",
             # Missing required variables
         }
-        
+
         result = await latex_tools.validate_document_data("Attack_Report", incomplete_data)
-        
+
         assert result["success"] is True
         assert result["validation"]["valid"] is False
         assert len(result["validation"]["errors"]) > 0
         assert len(result["validation"]["missing_variables"]) > 0
 
     @pytest.mark.asyncio
-    async def test_validate_document_data_unused_variables(self, latex_tools: LaTeXTemplateTools, sample_document_data: Dict[str, Any]) -> None:
+    async def test_validate_document_data_unused_variables(
+        self, latex_tools: LaTeXTemplateTools, sample_document_data: dict[str, Any]
+    ) -> None:
         """Test validation of document data with unused variables."""
         data_with_extras = sample_document_data.copy()
         data_with_extras["EXTRA_VARIABLE"] = "extra value"
-        
+
         result = await latex_tools.validate_document_data("Attack_Report", data_with_extras)
-        
+
         assert result["success"] is True
         assert result["validation"]["valid"] is True
         assert len(result["validation"]["warnings"]) > 0
         assert "EXTRA_VARIABLE" in result["validation"]["unused_variables"]
 
     @pytest.mark.asyncio
-    @patch('subprocess.run')
-    async def test_generate_document_pdf_success(self, mock_run: MagicMock, latex_tools: LaTeXTemplateTools, sample_document_data: Dict[str, Any]) -> None:
+    async def test_generate_document_pdf_success(
+        self,
+        latex_tools: LaTeXTemplateTools,
+        sample_document_data: dict[str, Any],
+    ) -> None:
         """Test successful PDF document generation."""
-        # Mock pdflatex availability check
-        mock_run.return_value.returncode = 0
-        
-        # Mock successful compilation
-        mock_run.return_value.stdout = "LaTeX compilation successful"
-        mock_run.return_value.stderr = ""
-        
-        # Patch Path.exists to simulate PDF existence only for the expected PDF path
-        original_exists = Path.exists
-        def fake_exists(self):
-            if self.name == "main_report.pdf":
-                # Actually create the file if it doesn't exist
-                if not original_exists(self):
-                    self.write_text("PDF content")
-                return True
-            return original_exists(self)
-        
-        with patch('pathlib.Path.exists', new=fake_exists):
-            result = await latex_tools.generate_document(
-                template_name="Attack_Report",
-                document_data=sample_document_data,
-                output_format="pdf"
-            )
-        
+        # Mock subprocess.run at the module level to work
+        # with our global fixture
+        with patch('src.latex_template_tools.subprocess.run') as mock_run:
+            # Mock pdflatex availability check
+            mock_result = MagicMock()
+            mock_result.returncode = 0
+            mock_result.stdout = "LaTeX compilation successful"
+            mock_result.stderr = ""
+            mock_run.return_value = mock_result
+
+            # Mock Path.exists to simulate PDF existence
+            with patch('pathlib.Path.exists', return_value=True):
+                result = await latex_tools.generate_document(
+                    template_name="Attack_Report",
+                    document_data=sample_document_data,
+                    output_format="pdf",
+                )
+
         assert result["success"] is True
         assert result["document"] is not None
         assert result["document"]["template_name"] == "Attack_Report"
         assert result["document"]["output_format"] == "pdf"
 
     @pytest.mark.asyncio
-    @patch('subprocess.run')
-    async def test_generate_document_pdf_compilation_failure(self, mock_run: MagicMock, latex_tools: LaTeXTemplateTools, sample_document_data: Dict[str, Any]) -> None:
+    async def test_generate_document_pdf_compilation_failure(
+        self,
+        latex_tools: LaTeXTemplateTools,
+        sample_document_data: dict[str, Any],
+    ) -> None:
         """Test PDF document generation with compilation failure."""
-        # Mock pdflatex availability check
-        mock_run.return_value.returncode = 0
-        
-        # Mock compilation failure
-        mock_run.return_value.stdout = ""
-        mock_run.return_value.stderr = "LaTeX compilation failed"
-        
-        # Patch Path.exists to simulate PDF not being created, but allow template files to be found
-        original_exists = Path.exists
-        def fake_exists(self):
-            if self.name == "main_report.pdf":
-                return False
-            return original_exists(self)
-        
-        with patch('pathlib.Path.exists', new=fake_exists):
-            result = await latex_tools.generate_document(
-                template_name="Attack_Report",
-                document_data=sample_document_data,
-                output_format="pdf"
-            )
-        
+        # Mock the subprocess.run at the module level to work with our global fixture
+        with patch('src.latex_template_tools.subprocess.run') as mock_run:
+            # Two returns: first availability check (success),
+            # then compilation (failure)
+            availability_result = MagicMock()
+            availability_result.returncode = 0  # pdflatex is available
+            availability_result.stdout = "/usr/bin/pdflatex"
+            availability_result.stderr = ""
+
+            compilation_result = MagicMock()
+            compilation_result.returncode = 1  # Compilation failure
+            compilation_result.stdout = ""
+            compilation_result.stderr = "LaTeX compilation failed"
+
+            mock_run.side_effect = [availability_result, compilation_result]
+
+            # Mock Path.exists: template files exist but final PDF
+            # does not (compilation failed)
+            def mock_exists(path_self):
+                # Allow template files and directories to exist, but not the final PDF
+                path_str = str(path_self)
+                if path_str.endswith('main_report.pdf') or path_str.endswith('.pdf'):
+                    return False  # PDF doesn't exist because compilation failed
+                return True  # Everything else exists (templates, directories, etc.)
+
+            with patch.object(Path, 'exists', mock_exists):
+                result = await latex_tools.generate_document(
+                    template_name="Attack_Report",
+                    document_data=sample_document_data,
+                    output_format="pdf",
+                )
+
         assert result["success"] is False
         assert "error" in result
         assert "PDF generation failed" in result["error"]
 
     @pytest.mark.asyncio
-    @patch('subprocess.run')
-    async def test_generate_document_pdflatex_not_found(self, mock_run: MagicMock, latex_tools: LaTeXTemplateTools, sample_document_data: Dict[str, Any]) -> None:
+    async def test_generate_document_pdflatex_not_found(
+        self,
+        latex_tools: LaTeXTemplateTools,
+        sample_document_data: dict[str, Any],
+    ) -> None:
         """Test PDF document generation when pdflatex is not available."""
-        # Mock pdflatex not found
-        mock_run.return_value.returncode = 1
-        
-        result = await latex_tools.generate_document(
-            template_name="Attack_Report",
-            document_data=sample_document_data,
-            output_format="pdf"
-        )
-        
+        # Mock the subprocess.run at the module level to simulate pdflatex not found
+        with patch('src.latex_template_tools.subprocess.run') as mock_run:
+            # Mock pdflatex not found (which command returns non-zero)
+            availability_result = MagicMock()
+            availability_result.returncode = 1  # pdflatex not found
+            availability_result.stdout = ""
+            availability_result.stderr = ""
+            mock_run.return_value = availability_result
+
+            result = await latex_tools.generate_document(
+                template_name="Attack_Report",
+                document_data=sample_document_data,
+                output_format="pdf",
+            )
+
         assert result["success"] is False
         assert "error" in result
         assert "pdflatex not found" in result["error"]
 
     @pytest.mark.asyncio
-    async def test_generate_document_tex_format(self, latex_tools: LaTeXTemplateTools, sample_document_data: Dict[str, Any]) -> None:
+    async def test_generate_document_tex_format(
+        self, latex_tools: LaTeXTemplateTools, sample_document_data: dict[str, Any]
+    ) -> None:
         """Test LaTeX document generation in TEX format."""
         result = await latex_tools.generate_document(
-            template_name="Attack_Report",
-            document_data=sample_document_data,
-            output_format="tex"
+            template_name="Attack_Report", document_data=sample_document_data, output_format="tex"
         )
-        
+
         assert result["success"] is True
         assert result["document"] is not None
         assert result["document"]["template_name"] == "Attack_Report"
@@ -311,14 +343,14 @@ Analysis of campaign {CAMPAIGN_NAME}.
         assert "generated_files" in result["document"]
 
     @pytest.mark.asyncio
-    async def test_generate_document_invalid_template(self, latex_tools: LaTeXTemplateTools, sample_document_data: Dict[str, Any]) -> None:
+    async def test_generate_document_invalid_template(
+        self, latex_tools: LaTeXTemplateTools, sample_document_data: dict[str, Any]
+    ) -> None:
         """Test document generation with invalid template."""
         result = await latex_tools.generate_document(
-            template_name="InvalidTemplate",
-            document_data=sample_document_data,
-            output_format="pdf"
+            template_name="InvalidTemplate", document_data=sample_document_data, output_format="pdf"
         )
-        
+
         assert result["success"] is False
         assert "error" in result
         assert "Template not found" in result["error"]
@@ -330,13 +362,11 @@ Analysis of campaign {CAMPAIGN_NAME}.
             "REPORT_TITLE": "Test Report"
             # Missing required variables
         }
-        
+
         result = await latex_tools.generate_document(
-            template_name="Attack_Report",
-            document_data=invalid_data,
-            output_format="pdf"
+            template_name="Attack_Report", document_data=invalid_data, output_format="pdf"
         )
-        
+
         assert result["success"] is False
         assert "error" in result
         assert "Invalid document data" in result["error"]
@@ -350,16 +380,16 @@ Analysis of campaign {CAMPAIGN_NAME}.
 \date{{{REPORT_DATE}}}
 Report Number: {{REPORT_NUMBER}}
 """
-        
+
         document_data = {
             "REPORT_TITLE": "Test Report",
             "AUTHOR_NAME": "Test Analyst",
             "REPORT_DATE": "2024-01-15",
-            "REPORT_NUMBER": "REP-001"
+            "REPORT_NUMBER": "REP-001",
         }
-        
+
         processed_content = latex_tools._substitute_variables(content, document_data)
-        
+
         assert "Test Report" in processed_content
         assert "Test Analyst" in processed_content
         assert "2024-01-15" in processed_content
@@ -374,9 +404,9 @@ Report Number: {{REPORT_NUMBER}}
                 "metadata": ["REPORT_DATE", "AUTHOR_NAME", "TOTAL_COUNT", "CONFIDENCE_SCORE"]
             }
         }
-        
+
         variable_types = latex_tools._infer_variable_types(template_config)
-        
+
         assert variable_types["REPORT_DATE"] == "datetime"
         assert variable_types["AUTHOR_NAME"] == "string"
         assert variable_types["TOTAL_COUNT"] == "integer"
@@ -390,9 +420,9 @@ Report Number: {{REPORT_NUMBER}}
                 "metadata": ["REPORT_DATE", "AUTHOR_NAME", "TOTAL_COUNT", "CONFIDENCE_SCORE"]
             }
         }
-        
+
         example_data = latex_tools._generate_example_data(template_config)
-        
+
         assert "REPORT_DATE" in example_data
         assert "AUTHOR_NAME" in example_data
         assert "TOTAL_COUNT" in example_data
@@ -405,7 +435,7 @@ Report Number: {{REPORT_NUMBER}}
         """Test section descriptions retrieval."""
         template_path = latex_tools.template_base_path
         descriptions = latex_tools._get_section_descriptions(template_path)
-        
+
         assert "title_page" in descriptions
         assert "executive_summary" in descriptions
         assert "campaign_overview" in descriptions
@@ -413,14 +443,16 @@ Report Number: {{REPORT_NUMBER}}
         assert len(descriptions["title_page"]) > 0
 
     @pytest.mark.asyncio
-    async def test_copy_template_files(self, latex_tools: LaTeXTemplateTools, tmp_path: Path) -> None:
+    async def test_copy_template_files(
+        self, latex_tools: LaTeXTemplateTools, tmp_path: Path
+    ) -> None:
         """Test template file copying."""
         source_path = latex_tools.template_base_path
         dest_path = tmp_path / "dest"
         dest_path.mkdir()
-        
+
         latex_tools._copy_template_files(source_path, dest_path, include_assets=True)
-        
+
         # Check that files were copied
         assert (dest_path / "main_report.tex").exists()
         assert (dest_path / "document_body.tex").exists()
@@ -429,7 +461,9 @@ Report Number: {{REPORT_NUMBER}}
         assert (dest_path / "sections" / "title_page.tex").exists()
 
     @pytest.mark.asyncio
-    async def test_process_template_file(self, latex_tools: LaTeXTemplateTools, tmp_path: Path) -> None:
+    async def test_process_template_file(
+        self, latex_tools: LaTeXTemplateTools, tmp_path: Path
+    ) -> None:
         """Test template file processing with variable substitution."""
         # Create a test file
         test_file = tmp_path / "test.tex"
@@ -437,14 +471,11 @@ Report Number: {{REPORT_NUMBER}}
 \title{{{REPORT_TITLE}}}
 \author{{{AUTHOR_NAME}}}
 """)
-        
-        document_data = {
-            "REPORT_TITLE": "Test Report",
-            "AUTHOR_NAME": "Test Analyst"
-        }
-        
+
+        document_data = {"REPORT_TITLE": "Test Report", "AUTHOR_NAME": "Test Analyst"}
+
         await latex_tools._process_template_file(test_file, document_data)
-        
+
         # Check that variables were substituted
         content = test_file.read_text()
         assert "Test Report" in content
@@ -458,31 +489,31 @@ Report Number: {{REPORT_NUMBER}}
         # Create temporary directory with output files
         temp_path = tmp_path / "temp"
         temp_path.mkdir()
-        
+
         # Create mock output files
         (temp_path / "main_report.pdf").write_text("PDF content")
         (temp_path / "main_report.tex").write_text("TEX content")
-        
+
         output_files = latex_tools._copy_output_files(temp_path, "Attack_Report")
-        
+
         # Check that output files were created
         assert "pdf" in output_files
         assert "tex" in output_files
-        
+
         # Check that files exist in output directory
         pdf_path = Path(output_files["pdf"])
         tex_path = Path(output_files["tex"])
-        
+
         assert pdf_path.exists()
         assert tex_path.exists()
         assert pdf_path.read_text() == "PDF content"
-        assert tex_path.read_text() == "TEX content" 
+        assert tex_path.read_text() == "TEX content"
 
     def test_find_project_root_from_src_directory(self) -> None:
         """Test that project root is found correctly from src directory."""
         tools = LaTeXTemplateTools()
         project_root = tools._find_project_root()
-        
+
         # Should find the project root (where setup.py exists)
         assert (project_root / "setup.py").exists()
         assert project_root.name == "dshield-mcp"
@@ -490,7 +521,7 @@ Report Number: {{REPORT_NUMBER}}
     def test_template_path_resolution(self) -> None:
         """Test that template path is resolved correctly relative to project root."""
         tools = LaTeXTemplateTools()
-        
+
         # Should resolve to project_root/templates/Attack_Report
         expected_path = tools._find_project_root() / "templates" / "Attack_Report"
         assert tools.template_base_path == expected_path
@@ -500,7 +531,7 @@ Report Number: {{REPORT_NUMBER}}
         """Test that custom template path is used when provided."""
         custom_path = Path("/tmp/custom_templates")
         custom_path.mkdir(exist_ok=True)
-        
+
         try:
             tools = LaTeXTemplateTools(str(custom_path))
             assert tools.template_base_path == custom_path
@@ -523,26 +554,26 @@ Report Number: {{REPORT_NUMBER}}
     def test_initialization_from_different_working_directories(self) -> None:
         """Test that initialization works from different working directories."""
         original_cwd = Path.cwd()
-        
+
         try:
             # Test from project root
             os.chdir(Path(__file__).parent.parent)
             tools1 = LaTeXTemplateTools()
-            
+
             # Test from src directory
             os.chdir(Path(__file__).parent.parent / "src")
             tools2 = LaTeXTemplateTools()
-            
+
             # Test from templates directory
             os.chdir(Path(__file__).parent.parent / "templates")
             tools3 = LaTeXTemplateTools()
-            
+
             # All should resolve to the same template path
             expected_path = Path(__file__).parent.parent / "templates" / "Attack_Report"
             assert tools1.template_base_path == expected_path
             assert tools2.template_base_path == expected_path
             assert tools3.template_base_path == expected_path
-            
+
         finally:
             # Restore original working directory
             os.chdir(original_cwd)
@@ -550,144 +581,167 @@ Report Number: {{REPORT_NUMBER}}
 
 class TestLaTeXTemplateToolsErrorHandling:
     """Test LaTeX tools error handling with MCPErrorHandler."""
-    
+
     @pytest_asyncio.fixture
     async def latex_tools_with_error_handler(self, tmp_path: Path) -> LaTeXTemplateTools:
-        """Create LaTeXTemplateTools instance with error handler and temporary template directory."""
+        """Create LaTeXTemplateTools instance with error handler and temporary template
+        directory."""
         template_dir = tmp_path / "templates" / "Attack_Report"
         template_dir.mkdir(parents=True)
-        
+
         # Create minimal template_info.json
         template_config = {
             "template_name": "Test Template",
             "version": "1.0",
             "description": "Test template for error handling",
             "sections": ["title_page"],
-            "required_variables": {
-                "metadata": ["REPORT_TITLE", "AUTHOR_NAME"]
-            }
+            "required_variables": {"metadata": ["REPORT_TITLE", "AUTHOR_NAME"]},
         }
-        
+
         with open(template_dir / "template_info.json", 'w') as f:
             json.dump(template_config, f)
-        
+
         # Create error handler
         from src.mcp_error_handler import MCPErrorHandler
+
         error_handler = MCPErrorHandler()
-        
+
         return LaTeXTemplateTools(str(template_dir), error_handler=error_handler)
-    
+
     @pytest_asyncio.fixture
     async def latex_tools_without_error_handler(self, tmp_path: Path) -> LaTeXTemplateTools:
         """Create LaTeXTemplateTools instance without error handler."""
         template_dir = tmp_path / "templates" / "Attack_Report"
         template_dir.mkdir(parents=True)
-        
+
         # Create minimal template_info.json
         template_config = {
             "template_name": "Test Template",
             "version": "1.0",
             "description": "Test template for error handling",
             "sections": ["title_page"],
-            "required_variables": {
-                "metadata": ["REPORT_TITLE", "AUTHOR_NAME"]
-            }
+            "required_variables": {"metadata": ["REPORT_TITLE", "AUTHOR_NAME"]},
         }
-        
+
         with open(template_dir / "template_info.json", 'w') as f:
             json.dump(template_config, f)
-        
+
         return LaTeXTemplateTools(str(template_dir))
-    
-    def test_init_with_error_handler(self, latex_tools_with_error_handler: LaTeXTemplateTools) -> None:
+
+    def test_init_with_error_handler(
+        self, latex_tools_with_error_handler: LaTeXTemplateTools
+    ) -> None:
         """Test LaTeXTemplateTools initialization with error handler."""
         assert latex_tools_with_error_handler.error_handler is not None
         assert hasattr(latex_tools_with_error_handler.error_handler, 'create_internal_error')
-    
-    def test_init_without_error_handler(self, latex_tools_without_error_handler: LaTeXTemplateTools) -> None:
+
+    def test_init_without_error_handler(
+        self, latex_tools_without_error_handler: LaTeXTemplateTools
+    ) -> None:
         """Test LaTeXTemplateTools initialization without error handler."""
         assert latex_tools_without_error_handler.error_handler is None
-    
+
     @pytest.mark.asyncio
-    async def test_generate_document_with_error_handler_template_not_found(self, latex_tools_with_error_handler: LaTeXTemplateTools) -> None:
+    async def test_generate_document_with_error_handler_template_not_found(
+        self, latex_tools_with_error_handler: LaTeXTemplateTools
+    ) -> None:
         """Test generate_document with error handler when template not found."""
         result = await latex_tools_with_error_handler.generate_document(
-            "NonExistentTemplate",
-            {"REPORT_TITLE": "Test", "AUTHOR_NAME": "Test"}
+            "NonExistentTemplate", {"REPORT_TITLE": "Test", "AUTHOR_NAME": "Test"}
         )
-        
+
         # Should return error response instead of success=False
         assert "error" in result
-        assert result["error"]["error"]["code"] == latex_tools_with_error_handler.error_handler.INTERNAL_ERROR
-    
+        assert (
+            result["error"]["error"]["code"]
+            == latex_tools_with_error_handler.error_handler.INTERNAL_ERROR
+        )
+
     @pytest.mark.asyncio
-    async def test_generate_document_without_error_handler_template_not_found(self, latex_tools_without_error_handler: LaTeXTemplateTools) -> None:
+    async def test_generate_document_without_error_handler_template_not_found(
+        self, latex_tools_without_error_handler: LaTeXTemplateTools
+    ) -> None:
         """Test generate_document without error handler when template not found."""
         result = await latex_tools_without_error_handler.generate_document(
-            "NonExistentTemplate",
-            {"REPORT_TITLE": "Test", "AUTHOR_NAME": "Test"}
+            "NonExistentTemplate", {"REPORT_TITLE": "Test", "AUTHOR_NAME": "Test"}
         )
-        
+
         # Should return traditional error format
         assert result["success"] is False
         assert "error" in result
         assert "Template not found" in result["error"]
-    
+
     @pytest.mark.asyncio
-    async def test_list_available_templates_with_error_handler(self, latex_tools_with_error_handler: LaTeXTemplateTools, tmp_path: Path) -> None:
+    async def test_list_available_templates_with_error_handler(
+        self, latex_tools_with_error_handler: LaTeXTemplateTools, tmp_path: Path
+    ) -> None:
         """Test list_available_templates with error handler."""
         # Test that the method works correctly with error handler integration
         # The method is designed to be fault-tolerant, so we'll test that it handles
         # errors gracefully and continues to work
-        
+
         result = await latex_tools_with_error_handler.list_available_templates()
-        
+
         # Should return successful result
         assert result["success"] is True
         assert "templates" in result
         assert "total_templates" in result
-        
+
         # Verify that error handler is properly integrated
         assert latex_tools_with_error_handler.error_handler is not None
         assert hasattr(latex_tools_with_error_handler.error_handler, 'create_internal_error')
-    
+
     @pytest.mark.asyncio
-    async def test_get_template_schema_with_error_handler(self, latex_tools_with_error_handler: LaTeXTemplateTools) -> None:
+    async def test_get_template_schema_with_error_handler(
+        self, latex_tools_with_error_handler: LaTeXTemplateTools
+    ) -> None:
         """Test get_template_schema with error handler when template not found."""
         result = await latex_tools_with_error_handler.get_template_schema("NonExistentTemplate")
-        
+
         # Should return error response
         assert "error" in result
-        assert result["error"]["error"]["code"] == latex_tools_with_error_handler.error_handler.INTERNAL_ERROR
-    
+        assert (
+            result["error"]["error"]["code"]
+            == latex_tools_with_error_handler.error_handler.INTERNAL_ERROR
+        )
+
     @pytest.mark.asyncio
-    async def test_validate_document_data_with_error_handler(self, latex_tools_with_error_handler: LaTeXTemplateTools) -> None:
+    async def test_validate_document_data_with_error_handler(
+        self, latex_tools_with_error_handler: LaTeXTemplateTools
+    ) -> None:
         """Test validate_document_data with error handler when template not found."""
         result = await latex_tools_with_error_handler.validate_document_data(
-            "NonExistentTemplate",
-            {"REPORT_TITLE": "Test"}
+            "NonExistentTemplate", {"REPORT_TITLE": "Test"}
         )
-        
+
         # Should return error response
         assert "error" in result
-        assert result["error"]["error"]["code"] == latex_tools_with_error_handler.error_handler.INTERNAL_ERROR
-    
+        assert (
+            result["error"]["error"]["code"]
+            == latex_tools_with_error_handler.error_handler.INTERNAL_ERROR
+        )
+
     @pytest.mark.asyncio
-    async def test_compile_latex_document_with_error_handler(self, latex_tools_with_error_handler: LaTeXTemplateTools, tmp_path: Path) -> None:
+    async def test_compile_latex_document_with_error_handler(
+        self, latex_tools_with_error_handler: LaTeXTemplateTools, tmp_path: Path
+    ) -> None:
         """Test _compile_latex_document with error handler."""
         # Create a temporary directory with invalid LaTeX content
         temp_path = tmp_path / "temp"
         temp_path.mkdir()
-        
+
         # Create invalid LaTeX file
         (temp_path / "main_report.tex").write_text("\\invalid\\command")
-        
+
         # Mock subprocess to raise an exception
         with patch('subprocess.run') as mock_run:
             mock_run.side_effect = Exception("Test compilation error")
-            
+
             result = await latex_tools_with_error_handler._compile_latex_document(temp_path, {})
-        
+
         # Should return error response
         assert "error" in result
-        assert result["error"]["error"]["code"] == latex_tools_with_error_handler.error_handler.INTERNAL_ERROR 
+        assert (
+            result["error"]["error"]["code"]
+            == latex_tools_with_error_handler.error_handler.INTERNAL_ERROR
+        )
