@@ -1,4 +1,4 @@
 """Test suite for DShield MCP Elastic SIEM integration."""
 
 # This file is intentionally left mostly empty to mark the directory as a Python package.
-# All test modules are imported and run by pytest automatically. 
+# All test modules are imported and run by pytest automatically.
