@@ -1,14 +1,22 @@
 """Tests for data parsing and field mapping functionality."""
 
+from unittest.mock import Mock, patch
+
 import pytest
-import json
-from datetime import datetime, timedelta
-from unittest.mock import Mock, patch, AsyncMock
-from typing import Dict, Any, List
 
 from src.data_processor import DataProcessor
 from src.elasticsearch_client import ElasticsearchClient
 from src.models import DShieldAttack, EventSeverity
+
+
+# Fixture to mock get_user_config for all tests
+@pytest.fixture(autouse=True)
+def mock_user_config():
+    """Mock get_user_config to prevent subprocess calls."""
+    with patch('src.elasticsearch_client.get_user_config') as mock_get_user_config:
+        mock_user_config = Mock()
+        mock_get_user_config.return_value = mock_user_config
+        yield mock_user_config
 
 
 class MockUserConfig:
@@ -16,7 +24,7 @@ class MockUserConfig:
 
     def get_setting(self, *args, **kwargs):
         """Return None for any configuration setting request (mock behavior).
-        
+
         Returns:
             None: Always returns None for any setting.
 
@@ -26,11 +34,11 @@ class MockUserConfig:
 
 class TestDataProcessor:
     """Test DataProcessor functionality for parsing and processing security events."""
-    
+
     def setup_method(self):
         """Set up test fixtures."""
         self.data_processor = DataProcessor()
-        
+
         # Sample test events
         self.sample_events = [
             {
@@ -51,7 +59,7 @@ class TestDataProcessor:
                 'reputation_score': 85,
                 'attack_count': 5,
                 'tags': ['brute_force', 'suspicious'],
-                'attack_types': ['brute_force']
+                'attack_types': ['brute_force'],
             },
             {
                 'id': 'test-2',
@@ -71,10 +79,10 @@ class TestDataProcessor:
                 'reputation_score': 75,
                 'attack_count': 3,
                 'tags': ['port_scan', 'reconnaissance'],
-                'attack_types': ['port_scan']
-            }
+                'attack_types': ['port_scan'],
+            },
         ]
-        
+
         # Sample attack events
         self.sample_attacks = [
             {
@@ -98,17 +106,17 @@ class TestDataProcessor:
                 'tags': ['brute_force', 'ssh'],
                 'attack_types': ['brute_force'],
                 'raw_data': {'ssh_attempts': 50},
-                'indices': ['cowrie.dshield-2024.01.01']
+                'indices': ['cowrie.dshield-2024.01.01'],
             }
         ]
-    
+
     def test_process_security_events(self):
         """Test processing of security events."""
         processed_events = self.data_processor.process_security_events(self.sample_events)
-        
+
         assert len(processed_events) == 2
         assert all(isinstance(event, dict) for event in processed_events)
-        
+
         # Check first event processing
         first_event = processed_events[0]
         assert first_event['source_ip'] == '192.168.1.100'
@@ -116,40 +124,40 @@ class TestDataProcessor:
         assert first_event['event_type'] == 'authentication_failure'
         assert first_event['severity'] == 'high'
         assert 'attack_patterns' in first_event
-        
+
         # Check second event processing
         second_event = processed_events[1]
         assert second_event['source_ip'] == '192.168.1.101'
         assert second_event['event_type'] == 'port_scan'
         assert second_event['severity'] == 'medium'
-    
+
     def test_process_security_events_empty(self):
         """Test processing of empty event list."""
         processed_events = self.data_processor.process_security_events([])
         assert processed_events == []
-    
+
     def test_process_security_events_with_invalid_events(self):
         """Test processing with invalid events (should handle gracefully)."""
         invalid_events = [
             None,
             {},
             {'invalid': 'data'},
-            self.sample_events[0]  # One valid event
+            self.sample_events[0],  # One valid event
         ]
-        
+
         processed_events = self.data_processor.process_security_events(invalid_events)
-        
+
         # Should process the valid event and skip invalid ones
         assert len(processed_events) >= 1
         assert all(isinstance(event, dict) for event in processed_events)
-    
+
     def test_process_dshield_attacks(self):
         """Test processing of DShield attack events."""
         processed_attacks = self.data_processor.process_dshield_attacks(self.sample_attacks)
-        
+
         assert len(processed_attacks) == 1
         assert all(isinstance(attack, DShieldAttack) for attack in processed_attacks)
-        
+
         # Check attack processing
         attack = processed_attacks[0]
         assert attack.source_ip == '192.168.1.100'
@@ -160,16 +168,16 @@ class TestDataProcessor:
         assert attack.attack_count == 10
         assert 'brute_force' in attack.tags
         assert 'ssh' in attack.tags
-    
+
     def test_process_dshield_attacks_empty(self):
         """Test processing of empty attack list."""
         processed_attacks = self.data_processor.process_dshield_attacks([])
         assert processed_attacks == []
-    
+
     def test_generate_security_summary(self):
         """Test generation of security summary."""
         summary = self.data_processor.generate_security_summary(self.sample_events)
-        
+
         # Check basic summary structure
         assert 'timestamp' in summary
         assert 'total_events' in summary
@@ -180,7 +188,7 @@ class TestDataProcessor:
         assert 'top_source_ips' in summary
         assert 'high_risk_events' in summary
         assert 'attack_patterns' in summary
-        
+
         # Check summary values
         assert summary['total_events'] == 2
         assert summary['unique_source_ips'] == 2
@@ -190,18 +198,18 @@ class TestDataProcessor:
         assert 'medium' in summary['events_by_severity']
         assert summary['events_by_severity']['high'] == 1
         assert summary['events_by_severity']['medium'] == 1
-    
+
     def test_generate_security_summary_empty(self):
         """Test generation of security summary for empty events."""
         summary = self.data_processor.generate_security_summary([])
-        
+
         assert summary['total_events'] == 0
         assert summary['unique_source_ips'] == 0
         assert summary['unique_destination_ips'] == 0
         assert summary['high_risk_events'] == 0
         assert summary['events_by_severity'] == {}
         assert summary['events_by_category'] == {}
-    
+
     def test_generate_security_summary_with_dshield_data(self):
         """Test security summary with DShield-specific data."""
         dshield_events = [
@@ -218,7 +226,7 @@ class TestDataProcessor:
                 'attack_count': 15,
                 'country': 'US',
                 'asn': 'AS12345',
-                'organization': 'Test Corp'
+                'organization': 'Test Corp',
             },
             {
                 'id': 'dshield-2',
@@ -233,12 +241,12 @@ class TestDataProcessor:
                 'attack_count': 5,
                 'country': 'CN',
                 'asn': 'AS67890',
-                'organization': 'Unknown'
-            }
+                'organization': 'Unknown',
+            },
         ]
-        
+
         summary = self.data_processor.generate_security_summary(dshield_events)
-        
+
         # Check DShield-specific fields
         assert 'dshield_attacks' in summary
         assert 'dshield_blocks' in summary
@@ -248,13 +256,13 @@ class TestDataProcessor:
         assert 'asn_distribution' in summary
         assert 'organization_distribution' in summary
         assert 'reputation_distribution' in summary
-        
+
         # Check DShield values
         assert summary['dshield_attacks'] == 1
         assert summary['dshield_blocks'] == 1
         # Note: reputation_hits logic may vary based on implementation
         assert summary['dshield_reputation_hits'] >= 0
-    
+
     def test_attack_pattern_detection(self):
         """Test detection of attack patterns in events."""
         pattern_events = [
@@ -267,7 +275,7 @@ class TestDataProcessor:
                 'severity': 'high',
                 'category': 'authentication',
                 'description': 'Failed login attempt - brute force attack detected',
-                'tags': ['brute_force', 'failed_login']
+                'tags': ['brute_force', 'failed_login'],
             },
             {
                 'id': 'pattern-2',
@@ -278,43 +286,43 @@ class TestDataProcessor:
                 'severity': 'medium',
                 'category': 'reconnaissance',
                 'description': 'Port scan detected using nmap',
-                'tags': ['port_scan', 'nmap']
-            }
+                'tags': ['port_scan', 'nmap'],
+            },
         ]
-        
+
         processed_events = self.data_processor.process_security_events(pattern_events)
-        
+
         # Check that attack patterns are detected
         for event in processed_events:
             assert 'attack_patterns' in event
             assert isinstance(event['attack_patterns'], dict)
-    
+
     def test_extract_unique_ips(self):
         """Test extraction of unique IP addresses from events."""
         unique_ips = self.data_processor.extract_unique_ips(self.sample_events)
-        
+
         # extract_unique_ips returns both source and destination IPs
         assert len(unique_ips) == 4  # 2 source + 2 destination IPs
         assert '192.168.1.100' in unique_ips
         assert '192.168.1.101' in unique_ips
         assert '10.0.0.1' in unique_ips
         assert '10.0.0.2' in unique_ips
-    
+
     def test_extract_unique_ips_empty(self):
         """Test extraction of unique IPs from empty events."""
         unique_ips = self.data_processor.extract_unique_ips([])
         assert unique_ips == []
-    
+
     def test_extract_unique_ips_with_duplicates(self):
         """Test extraction of unique IPs with duplicate IPs."""
         duplicate_events = [
             {'source_ip': '192.168.1.100', 'destination_ip': '10.0.0.1'},
             {'source_ip': '192.168.1.100', 'destination_ip': '10.0.0.2'},
-            {'source_ip': '192.168.1.101', 'destination_ip': '10.0.0.1'}
+            {'source_ip': '192.168.1.101', 'destination_ip': '10.0.0.1'},
         ]
-        
+
         unique_ips = self.data_processor.extract_unique_ips(duplicate_events)
-        
+
         # Should return unique IPs from both source and destination
         assert len(unique_ips) == 4  # 2 unique source + 2 unique destination IPs
         assert '192.168.1.100' in unique_ips
@@ -325,35 +333,106 @@ class TestDataProcessor:
 
 class TestElasticsearchClientFieldMapping:
     """Test ElasticsearchClient field mapping functionality."""
-    
+
     def setup_method(self):
         """Set up test fixtures."""
         # Create a mock client without real dependencies
         self.client = Mock()
-        
+
         # Mock the field mappings
         self.client.dshield_field_mappings = {
             'timestamp': ['@timestamp', 'timestamp', 'time', 'date', 'event.ingested'],
             'source_ip': [
-                'source.ip', 'src_ip', 'srcip', 'sourceip', 'attacker_ip', 'attackerip', 'src', 'client_ip', 'ip.src', 'ip_source', 'source.address', 'related.ip'
+                'source.ip',
+                'src_ip',
+                'srcip',
+                'sourceip',
+                'attacker_ip',
+                'attackerip',
+                'src',
+                'client_ip',
+                'ip.src',
+                'ip_source',
+                'source.address',
+                'related.ip',
             ],
             'destination_ip': [
-                'destination.ip', 'dst_ip', 'dstip', 'destinationip', 'target_ip', 'targetip', 'dst', 'server_ip', 'ip.dst', 'ip_destination', 'destination.address', 'related.ip'
+                'destination.ip',
+                'dst_ip',
+                'dstip',
+                'destinationip',
+                'target_ip',
+                'targetip',
+                'dst',
+                'server_ip',
+                'ip.dst',
+                'ip_destination',
+                'destination.address',
+                'related.ip',
             ],
             'source_port': [
-                'source.port', 'src_port', 'srcport', 'sourceport', 'attacker_port', 'sport', 'client_port', 'port.src', 'port_source'
+                'source.port',
+                'src_port',
+                'srcport',
+                'sourceport',
+                'attacker_port',
+                'sport',
+                'client_port',
+                'port.src',
+                'port_source',
             ],
             'destination_port': [
-                'destination.port', 'dst_port', 'dstport', 'destinationport', 'target_port', 'dport', 'server_port', 'port.dst', 'port_destination'
+                'destination.port',
+                'dst_port',
+                'dstport',
+                'destinationport',
+                'target_port',
+                'dport',
+                'server_port',
+                'port.dst',
+                'port_destination',
             ],
             'event_type': ['event.type', 'type', 'eventtype', 'event_type', 'event.category'],
             'category': ['event.category', 'category', 'eventcategory', 'event_category'],
-            'severity': ['event.severity', 'severity', 'level', 'risk_level', 'threat_level', 'event.level'],
-            'description': ['event.description', 'message', 'description', 'summary', 'attack_description', 'event.original'],
-            'protocol': ['network.protocol', 'protocol', 'proto', 'transport_protocol', 'event.protocol', 'ip.proto'],
-            'country': ['source.geo.country_name', 'country', 'country_name', 'geo.country', 'source.country'],
+            'severity': [
+                'event.severity',
+                'severity',
+                'level',
+                'risk_level',
+                'threat_level',
+                'event.level',
+            ],
+            'description': [
+                'event.description',
+                'message',
+                'description',
+                'summary',
+                'attack_description',
+                'event.original',
+            ],
+            'protocol': [
+                'network.protocol',
+                'protocol',
+                'proto',
+                'transport_protocol',
+                'event.protocol',
+                'ip.proto',
+            ],
+            'country': [
+                'source.geo.country_name',
+                'country',
+                'country_name',
+                'geo.country',
+                'source.country',
+            ],
             'asn': ['asn', 'as_number', 'autonomous_system', 'attacker_asn', 'source.geo.asn'],
-            'organization': ['org', 'organization', 'org_name', 'attacker_org', 'source.geo.organization_name'],
+            'organization': [
+                'org',
+                'organization',
+                'org_name',
+                'attacker_org',
+                'source.geo.organization_name',
+            ],
             'reputation_score': ['reputation', 'reputation_score', 'dshield_score', 'threat_score'],
             'attack_count': ['count', 'attack_count', 'hits', 'attempts'],
             'first_seen': ['firstseen', 'first_seen', 'first_seen_date'],
@@ -361,7 +440,7 @@ class TestElasticsearchClientFieldMapping:
             'tags': ['tags', 'event.tags', 'labels', 'categories'],
             'attack_types': ['attacks', 'attack_types', 'attack_methods'],
         }
-        
+
         # Mock the _extract_field_mapped method
         def mock_extract_field_mapped(source, field_type, default=None):
             if field_type in self.client.dshield_field_mappings:
@@ -383,14 +462,14 @@ class TestElasticsearchClientFieldMapping:
                         if value is not None:
                             return value
             return default
-        
+
         self.client._extract_field_mapped = mock_extract_field_mapped
-        
+
         # Mock the _map_query_fields method
         def mock_map_query_fields(filters):
             if not filters:
                 return {}
-                
+
             field_mappings = {
                 "source_ip": "source.ip",
                 "src_ip": "source.ip",
@@ -444,177 +523,153 @@ class TestElasticsearchClientFieldMapping:
                 "severity": "event.severity",
                 "description": "event.description",
                 "message": "log.message",
-                "log_message": "log.message"
+                "log_message": "log.message",
             }
-            
+
             mapped_filters = {}
             for key, value in filters.items():
                 mapped_key = field_mappings.get(key, key)
                 mapped_filters[mapped_key] = value
-            
+
             return mapped_filters
-        
+
         self.client._map_query_fields = mock_map_query_fields
-    
+
     def test_extract_field_mapped_basic(self):
         """Test basic field mapping extraction."""
         source_data = {
             "source.ip": "192.168.1.100",
             "destination.ip": "10.0.0.1",
             "event.type": "authentication_failure",
-            "event.severity": "high"
+            "event.severity": "high",
         }
-        
+
         # Test source_ip mapping
         source_ip = self.client._extract_field_mapped(source_data, "source_ip")
         assert source_ip == "192.168.1.100"
-        
+
         # Test destination_ip mapping
         dest_ip = self.client._extract_field_mapped(source_data, "destination_ip")
         assert dest_ip == "10.0.0.1"
-        
+
         # Test event_type mapping
         event_type = self.client._extract_field_mapped(source_data, "event_type")
         assert event_type == "authentication_failure"
-        
+
         # Test severity mapping
         severity = self.client._extract_field_mapped(source_data, "severity")
         assert severity == "high"
-    
+
     def test_extract_field_mapped_alternative_fields(self):
         """Test field mapping with alternative field names."""
         source_data = {
             "src_ip": "192.168.1.100",
             "dst_ip": "10.0.0.1",
             "type": "port_scan",
-            "level": "medium"
+            "level": "medium",
         }
-        
+
         # Test alternative source_ip fields
         source_ip = self.client._extract_field_mapped(source_data, "source_ip")
         assert source_ip == "192.168.1.100"
-        
+
         # Test alternative destination_ip fields
         dest_ip = self.client._extract_field_mapped(source_data, "destination_ip")
         assert dest_ip == "10.0.0.1"
-        
+
         # Test alternative event_type fields
         event_type = self.client._extract_field_mapped(source_data, "event_type")
         assert event_type == "port_scan"
-        
+
         # Test alternative severity fields
         severity = self.client._extract_field_mapped(source_data, "severity")
         assert severity == "medium"
-    
+
     def test_extract_field_mapped_nested_fields(self):
         """Test field mapping with nested field structures."""
         source_data = {
-            "source": {
-                "ip": "192.168.1.100",
-                "port": 12345,
-                "geo": {
-                    "country_name": "US"
-                }
-            },
-            "destination": {
-                "ip": "10.0.0.1",
-                "port": 80
-            },
+            "source": {"ip": "192.168.1.100", "port": 12345, "geo": {"country_name": "US"}},
+            "destination": {"ip": "10.0.0.1", "port": 80},
             "event": {
                 "type": "authentication_failure",
                 "severity": "high",
-                "description": "Failed login attempt"
-            }
+                "description": "Failed login attempt",
+            },
         }
-        
+
         # Test nested source_ip
         source_ip = self.client._extract_field_mapped(source_data, "source_ip")
         assert source_ip == "192.168.1.100"
-        
+
         # Test nested destination_ip
         dest_ip = self.client._extract_field_mapped(source_data, "destination_ip")
         assert dest_ip == "10.0.0.1"
-        
+
         # Test nested event_type
         event_type = self.client._extract_field_mapped(source_data, "event_type")
         assert event_type == "authentication_failure"
-        
+
         # Test nested severity
         severity = self.client._extract_field_mapped(source_data, "severity")
         assert severity == "high"
-        
+
         # Test nested country
         country = self.client._extract_field_mapped(source_data, "country")
         assert country == "US"
-    
+
     def test_extract_field_mapped_missing_fields(self):
         """Test field mapping with missing fields."""
-        source_data = {
-            "event.type": "authentication_failure"
-        }
-        
+        source_data = {"event.type": "authentication_failure"}
+
         # Test missing source_ip
         source_ip = self.client._extract_field_mapped(source_data, "source_ip")
         assert source_ip is None
-        
+
         # Test missing destination_ip with default
         dest_ip = self.client._extract_field_mapped(source_data, "destination_ip", "unknown")
         assert dest_ip == "unknown"
-        
+
         # Test missing severity with default
         severity = self.client._extract_field_mapped(source_data, "severity", "medium")
         assert severity == "medium"
-    
+
     def test_extract_field_mapped_complex_nested(self):
         """Test field mapping with complex nested structures."""
         source_data = {
             "source": {
                 "address": "192.168.1.100",
                 "port": 12345,
-                "geo": {
-                    "country_name": "US",
-                    "asn": "AS12345",
-                    "organization_name": "Test Corp"
-                }
+                "geo": {"country_name": "US", "asn": "AS12345", "organization_name": "Test Corp"},
             },
-            "destination": {
-                "address": "10.0.0.1",
-                "port": 80
-            },
+            "destination": {"address": "10.0.0.1", "port": 80},
             "event": {
                 "type": "authentication_failure",
                 "severity": "high",
                 "description": "Failed login attempt",
-                "category": "authentication"
+                "category": "authentication",
             },
-            "network": {
-                "protocol": "tcp",
-                "direction": "inbound"
-            },
-            "http": {
-                "request": {
-                    "method": "POST"
-                },
-                "response": {
-                    "status_code": 401
-                }
-            }
+            "network": {"protocol": "tcp", "direction": "inbound"},
+            "http": {"request": {"method": "POST"}, "response": {"status_code": 401}},
         }
-        
+
         # Test all field mappings
         assert self.client._extract_field_mapped(source_data, "source_ip") == "192.168.1.100"
         assert self.client._extract_field_mapped(source_data, "destination_ip") == "10.0.0.1"
         assert self.client._extract_field_mapped(source_data, "source_port") == 12345
         assert self.client._extract_field_mapped(source_data, "destination_port") == 80
-        assert self.client._extract_field_mapped(source_data, "event_type") == "authentication_failure"
+        assert (
+            self.client._extract_field_mapped(source_data, "event_type") == "authentication_failure"
+        )
         assert self.client._extract_field_mapped(source_data, "severity") == "high"
         assert self.client._extract_field_mapped(source_data, "category") == "authentication"
-        assert self.client._extract_field_mapped(source_data, "description") == "Failed login attempt"
+        assert (
+            self.client._extract_field_mapped(source_data, "description") == "Failed login attempt"
+        )
         assert self.client._extract_field_mapped(source_data, "protocol") == "tcp"
         assert self.client._extract_field_mapped(source_data, "country") == "US"
         assert self.client._extract_field_mapped(source_data, "asn") == "AS12345"
         assert self.client._extract_field_mapped(source_data, "organization") == "Test Corp"
-    
+
     def test_map_query_fields(self):
         """Test mapping of query fields from user-friendly names to ECS notation."""
         user_filters = {
@@ -622,86 +677,160 @@ class TestElasticsearchClientFieldMapping:
             "destination_ip": "10.0.0.1",
             "event_type": "authentication_failure",
             "severity": "high",
-            "protocol": "tcp"
+            "protocol": "tcp",
         }
-        
+
         mapped_filters = self.client._map_query_fields(user_filters)
-        
+
         # Check that fields are mapped to ECS notation
         assert mapped_filters["source.ip"] == "192.168.1.100"
         assert mapped_filters["destination.ip"] == "10.0.0.1"
         assert mapped_filters["event.type"] == "authentication_failure"
         assert mapped_filters["event.severity"] == "high"
         assert mapped_filters["network.protocol"] == "tcp"
-    
+
     def test_map_query_fields_with_arrays(self):
         """Test mapping of query fields with array values."""
         user_filters = {
             "source_ip": ["192.168.1.100", "192.168.1.101"],
             "event_type": ["authentication_failure", "port_scan"],
-            "severity": "high"
+            "severity": "high",
         }
-        
+
         mapped_filters = self.client._map_query_fields(user_filters)
-        
+
         assert mapped_filters["source.ip"] == ["192.168.1.100", "192.168.1.101"]
         assert mapped_filters["event.type"] == ["authentication_failure", "port_scan"]
         assert mapped_filters["event.severity"] == "high"
-    
+
     def test_map_query_fields_with_nested_filters(self):
         """Test mapping of query fields with nested filter structures."""
         user_filters = {
             "source_ip": {"eq": "192.168.1.100"},
             "timestamp": {"gte": "2024-01-01T00:00:00Z", "lte": "2024-01-01T23:59:59Z"},
-            "event_type": {"in": ["authentication_failure", "port_scan"]}
+            "event_type": {"in": ["authentication_failure", "port_scan"]},
         }
-        
+
         mapped_filters = self.client._map_query_fields(user_filters)
-        
+
         assert mapped_filters["source.ip"] == {"eq": "192.168.1.100"}
-        assert mapped_filters["@timestamp"] == {"gte": "2024-01-01T00:00:00Z", "lte": "2024-01-01T23:59:59Z"}
+        assert mapped_filters["@timestamp"] == {
+            "gte": "2024-01-01T00:00:00Z",
+            "lte": "2024-01-01T23:59:59Z",
+        }
         assert mapped_filters["event.type"] == {"in": ["authentication_failure", "port_scan"]}
-    
+
     def test_map_query_fields_empty(self):
         """Test mapping of empty query fields."""
         mapped_filters = self.client._map_query_fields({})
         assert mapped_filters == {}
-        
+
         mapped_filters = self.client._map_query_fields(None)
         assert mapped_filters == {}
 
 
 class TestEventParsing:
     """Test event parsing functionality."""
-    
+
     def setup_method(self):
         """Set up test fixtures."""
         # Create a mock client without real dependencies
         self.client = Mock()
-        
+
         # Mock the field mappings
         self.client.dshield_field_mappings = {
             'timestamp': ['@timestamp', 'timestamp', 'time', 'date', 'event.ingested'],
             'source_ip': [
-                'source.ip', 'src_ip', 'srcip', 'sourceip', 'attacker_ip', 'attackerip', 'src', 'client_ip', 'ip.src', 'ip_source', 'source.address', 'related.ip'
+                'source.ip',
+                'src_ip',
+                'srcip',
+                'sourceip',
+                'attacker_ip',
+                'attackerip',
+                'src',
+                'client_ip',
+                'ip.src',
+                'ip_source',
+                'source.address',
+                'related.ip',
             ],
             'destination_ip': [
-                'destination.ip', 'dst_ip', 'dstip', 'destinationip', 'target_ip', 'targetip', 'dst', 'server_ip', 'ip.dst', 'ip_destination', 'destination.address', 'related.ip'
+                'destination.ip',
+                'dst_ip',
+                'dstip',
+                'destinationip',
+                'target_ip',
+                'targetip',
+                'dst',
+                'server_ip',
+                'ip.dst',
+                'ip_destination',
+                'destination.address',
+                'related.ip',
             ],
             'source_port': [
-                'source.port', 'src_port', 'srcport', 'sourceport', 'attacker_port', 'sport', 'client_port', 'port.src', 'port_source'
+                'source.port',
+                'src_port',
+                'srcport',
+                'sourceport',
+                'attacker_port',
+                'sport',
+                'client_port',
+                'port.src',
+                'port_source',
             ],
             'destination_port': [
-                'destination.port', 'dst_port', 'dstport', 'destinationport', 'target_port', 'dport', 'server_port', 'port.dst', 'port_destination'
+                'destination.port',
+                'dst_port',
+                'dstport',
+                'destinationport',
+                'target_port',
+                'dport',
+                'server_port',
+                'port.dst',
+                'port_destination',
             ],
             'event_type': ['event.type', 'type', 'eventtype', 'event_type', 'event.category'],
             'category': ['event.category', 'category', 'eventcategory', 'event_category'],
-            'severity': ['event.severity', 'severity', 'level', 'risk_level', 'threat_level', 'event.level'],
-            'description': ['event.description', 'message', 'description', 'summary', 'attack_description', 'event.original'],
-            'protocol': ['network.protocol', 'protocol', 'proto', 'transport_protocol', 'event.protocol', 'ip.proto'],
-            'country': ['source.geo.country_name', 'country', 'country_name', 'geo.country', 'source.country'],
+            'severity': [
+                'event.severity',
+                'severity',
+                'level',
+                'risk_level',
+                'threat_level',
+                'event.level',
+            ],
+            'description': [
+                'event.description',
+                'message',
+                'description',
+                'summary',
+                'attack_description',
+                'event.original',
+            ],
+            'protocol': [
+                'network.protocol',
+                'protocol',
+                'proto',
+                'transport_protocol',
+                'event.protocol',
+                'ip.proto',
+            ],
+            'country': [
+                'source.geo.country_name',
+                'country',
+                'country_name',
+                'geo.country',
+                'source.country',
+            ],
             'asn': ['asn', 'as_number', 'autonomous_system', 'attacker_asn', 'source.geo.asn'],
-            'organization': ['org', 'organization', 'org_name', 'attacker_org', 'source.geo.organization_name'],
+            'organization': [
+                'org',
+                'organization',
+                'org_name',
+                'attacker_org',
+                'source.geo.organization_name',
+            ],
             'reputation_score': ['reputation', 'reputation_score', 'dshield_score', 'threat_score'],
             'attack_count': ['count', 'attack_count', 'hits', 'attempts'],
             'first_seen': ['firstseen', 'first_seen', 'first_seen_date'],
@@ -711,16 +840,16 @@ class TestEventParsing:
             'http_method': ['http.request.method', 'httpmethod', 'http_method'],
             'http_status': ['http.response.status_code', 'httpstatus', 'http_status'],
             'http_version': ['http.version', 'httpversion', 'http_version'],
-            'user_agent': ['user_agent.original', 'useragent', 'ua']
+            'user_agent': ['user_agent.original', 'useragent', 'ua'],
         }
-        
+
         # Mock the _parse_dshield_event method
         def mock_parse_dshield_event(hit, indices):
             try:
                 source = hit['_source']
                 if source is None:
                     return None
-                
+
                 # Extract basic fields
                 event = {
                     'id': hit['_id'],
@@ -744,26 +873,32 @@ class TestEventParsing:
                     'tags': source.get('tags', []),
                     'attack_types': source.get('attack_types', []),
                     'raw_data': source,
-                    'indices': indices
+                    'indices': indices,
                 }
-                
+
                 # Handle HTTP fields for description derivation
                 http_method = source.get('http.request.method') or source.get('http_method')
                 http_status = source.get('http.response.status_code') or source.get('http_status')
                 http_version = source.get('http.version') or source.get('http_version')
                 user_agent = source.get('user_agent.original') or source.get('user_agent')
-                
+
                 # Derive description from HTTP fields if not directly available
                 if not event['description']:
                     if http_method and http_status:
-                        event['description'] = f"HTTP {http_method} request with status {http_status}"
+                        event['description'] = (
+                            f"HTTP {http_method} request with status {http_status}"
+                        )
                     elif http_method:
                         event['description'] = f"HTTP {http_method} request"
                     elif user_agent:
                         event['description'] = f"Request with user agent: {user_agent[:50]}..."
                     else:
-                        event['description'] = f"{event['event_type']} event from {event['source_ip'] or 'unknown'} to {event['destination_ip'] or 'unknown'}"
-                
+                        event['description'] = (
+                            f"{event['event_type']} event from "
+                            f"{event['source_ip'] or 'unknown'} to "
+                            f"{event['destination_ip'] or 'unknown'}"
+                        )
+
                 # Derive protocol from HTTP fields if not directly available
                 if not event['protocol']:
                     if http_version:
@@ -774,7 +909,7 @@ class TestEventParsing:
                         event['protocol'] = 'http'
                     else:
                         event['protocol'] = 'unknown'
-                
+
                 # Provide fallback values for critical fields
                 if not event['severity']:
                     event['severity'] = 'medium'
@@ -782,14 +917,14 @@ class TestEventParsing:
                     event['category'] = 'other'
                 if not event['event_type']:
                     event['event_type'] = 'unknown'
-                
+
                 return event
-                
+
             except Exception:
                 return None
-        
+
         self.client._parse_dshield_event = mock_parse_dshield_event
-    
+
     def test_parse_dshield_event_basic(self):
         """Test basic DShield event parsing."""
         hit = {
@@ -807,13 +942,13 @@ class TestEventParsing:
                 'network.protocol': 'tcp',
                 'source.geo.country_name': 'US',
                 'asn': 'AS12345',
-                'organization': 'Test Corp'
-            }
+                'organization': 'Test Corp',
+            },
         }
-        
+
         indices = ['cowrie.dshield-2024.01.01']
         event = self.client._parse_dshield_event(hit, indices)
-        
+
         assert event is not None
         assert event['id'] == 'test-event-1'
         assert event['source_ip'] == '192.168.1.100'
@@ -829,7 +964,7 @@ class TestEventParsing:
         assert event['asn'] == 'AS12345'
         assert event['organization'] == 'Test Corp'
         assert event['indices'] == indices
-    
+
     def test_parse_dshield_event_with_http_fields(self):
         """Test DShield event parsing with HTTP fields."""
         hit = {
@@ -841,31 +976,32 @@ class TestEventParsing:
                 'http.request.method': 'POST',
                 'http.response.status_code': 401,
                 'http.version': '1.1',
-                'user_agent.original': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
-            }
+                'user_agent.original': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) '
+                'AppleWebKit/537.36',
+            },
         }
-        
+
         indices = ['zeek.dshield-2024.01.01']
         event = self.client._parse_dshield_event(hit, indices)
-        
+
         assert event is not None
         assert event['description'] == 'HTTP POST request with status 401'
         assert event['protocol'] == 'http'
-    
+
     def test_parse_dshield_event_with_missing_fields(self):
         """Test DShield event parsing with missing fields."""
         hit = {
             '_id': 'test-event-3',
             '_source': {
                 '@timestamp': '2024-01-01T12:00:00Z',
-                'source.ip': '192.168.1.100'
+                'source.ip': '192.168.1.100',
                 # Missing most fields
-            }
+            },
         }
-        
+
         indices = ['cowrie.dshield-2024.01.01']
         event = self.client._parse_dshield_event(hit, indices)
-        
+
         assert event is not None
         assert event['source_ip'] == '192.168.1.100'
         assert event['destination_ip'] is None
@@ -873,19 +1009,19 @@ class TestEventParsing:
         assert event['severity'] == 'medium'
         assert event['category'] == 'other'
         assert 'unknown' in event['description']
-    
+
     def test_parse_dshield_event_with_invalid_data(self):
         """Test DShield event parsing with invalid data."""
         hit = {
             '_id': 'test-event-4',
-            '_source': None  # Invalid source
+            '_source': None,  # Invalid source
         }
-        
+
         indices = ['cowrie.dshield-2024.01.01']
         event = self.client._parse_dshield_event(hit, indices)
-        
+
         assert event is None
-    
+
     def test_parse_dshield_event_with_alternative_field_names(self):
         """Test DShield event parsing with alternative field names."""
         hit = {
@@ -897,13 +1033,13 @@ class TestEventParsing:
                 'type': 'port_scan',
                 'level': 'medium',
                 'message': 'Port scan detected',
-                'proto': 'tcp'
-            }
+                'proto': 'tcp',
+            },
         }
-        
+
         indices = ['cowrie.dshield-2024.01.01']
         event = self.client._parse_dshield_event(hit, indices)
-        
+
         assert event is not None
         assert event['source_ip'] == '192.168.1.100'
         assert event['destination_ip'] == '10.0.0.1'
@@ -915,11 +1051,11 @@ class TestEventParsing:
 
 class TestDataProcessingWorkflow:
     """Test end-to-end data processing workflows."""
-    
+
     def setup_method(self):
         """Set up test fixtures."""
         self.data_processor = DataProcessor()
-        
+
         # Sample raw Elasticsearch hit
         self.raw_hit = {
             '_id': 'workflow-test-1',
@@ -940,14 +1076,16 @@ class TestDataProcessingWorkflow:
                 'reputation_score': 90,
                 'attack_count': 15,
                 'tags': ['brute_force', 'ssh'],
-                'attack_types': ['brute_force']
-            }
+                'attack_types': ['brute_force'],
+            },
         }
-    
+
     @patch('src.user_config.get_user_config')
     @patch('src.config_loader._resolve_secrets')
     @patch('src.config_loader.get_config')
-    def test_end_to_end_data_processing(self, mock_get_config, mock_resolve_secrets, mock_user_config):
+    def test_end_to_end_data_processing(
+        self, mock_get_config, mock_resolve_secrets, mock_user_config
+    ):
         """Test complete data processing workflow from raw hit to summary."""
         # Setup mocks
         mock_get_config.return_value = {
@@ -957,8 +1095,8 @@ class TestDataProcessingWorkflow:
                 'index_patterns': {
                     'cowrie': ['cowrie.dshield-*'],
                     'zeek': ['zeek.dshield-*'],
-                    'dshield': ['dshield-*']
-                }
+                    'dshield': ['dshield-*'],
+                },
             }
         }
         mock_resolve_secrets.return_value = {
@@ -968,11 +1106,11 @@ class TestDataProcessingWorkflow:
                 'index_patterns': {
                     'cowrie': ['cowrie.dshield-*'],
                     'zeek': ['zeek.dshield-*'],
-                    'dshield': ['dshield-*']
-                }
+                    'dshield': ['dshield-*'],
+                },
             }
         }
-        
+
         mock_user_config_instance = Mock()
         mock_user_config_instance.get_setting.side_effect = lambda section, key: {
             ("query", "default_page_size"): 100,
@@ -982,41 +1120,41 @@ class TestDataProcessingWorkflow:
             ("query", "enable_smart_optimization"): True,
             ("query", "fallback_strategy"): "aggregate",
             ("query", "max_query_complexity"): 10,
-            ("logging", "enable_performance_logging"): False
+            ("logging", "enable_performance_logging"): False,
         }.get((section, key), None)
         mock_user_config.return_value = mock_user_config_instance
-        
+
         client = ElasticsearchClient()
-        
+
         # Step 1: Parse raw Elasticsearch hit
         indices = ['cowrie.dshield-2024.01.01']
         parsed_event = client._parse_dshield_event(self.raw_hit, indices)
-        
+
         assert parsed_event is not None
         assert parsed_event['source_ip'] == '192.168.1.100'
         assert parsed_event['event_type'] == 'authentication_failure'
-        
+
         # Step 2: Process security events
         processed_events = self.data_processor.process_security_events([parsed_event])
-        
+
         assert len(processed_events) == 1
         assert 'attack_patterns' in processed_events[0]
-        
+
         # Step 3: Generate security summary
         summary = self.data_processor.generate_security_summary(processed_events)
-        
+
         assert summary['total_events'] == 1
         assert summary['unique_source_ips'] == 1
         assert summary['high_risk_events'] == 1
         assert 'high' in summary['events_by_severity']
-        
+
         # Step 4: Extract unique IPs
         unique_ips = self.data_processor.extract_unique_ips(processed_events)
-        
+
         assert len(unique_ips) == 2  # Both source and destination IPs
         assert '192.168.1.100' in unique_ips
         assert '10.0.0.1' in unique_ips
-    
+
     def test_data_processing_with_multiple_events(self):
         """Test data processing workflow with multiple events."""
         # Create multiple test events
@@ -1031,7 +1169,7 @@ class TestDataProcessingWorkflow:
                 'category': 'authentication',
                 'description': 'Failed login attempt',
                 'reputation_score': 90,
-                'attack_count': 10
+                'attack_count': 10,
             },
             {
                 'id': 'multi-2',
@@ -1043,7 +1181,7 @@ class TestDataProcessingWorkflow:
                 'category': 'reconnaissance',
                 'description': 'Port scan detected',
                 'reputation_score': 75,
-                'attack_count': 5
+                'attack_count': 5,
             },
             {
                 'id': 'multi-3',
@@ -1055,18 +1193,18 @@ class TestDataProcessingWorkflow:
                 'category': 'attack',
                 'description': 'SQL injection attempt',
                 'reputation_score': 95,
-                'attack_count': 20
-            }
+                'attack_count': 20,
+            },
         ]
-        
+
         # Process events
         processed_events = self.data_processor.process_security_events(test_events)
-        
+
         assert len(processed_events) == 3
-        
+
         # Generate summary
         summary = self.data_processor.generate_security_summary(processed_events)
-        
+
         assert summary['total_events'] == 3
         assert summary['unique_source_ips'] == 2  # Two unique source IPs
         assert summary['unique_destination_ips'] == 3  # Three unique destination IPs
@@ -1074,13 +1212,13 @@ class TestDataProcessingWorkflow:
         assert summary['events_by_severity']['high'] == 1
         assert summary['events_by_severity']['medium'] == 1
         assert summary['events_by_severity']['critical'] == 1
-        
+
         # Extract unique IPs
         unique_ips = self.data_processor.extract_unique_ips(processed_events)
-        
+
         assert len(unique_ips) == 5  # 2 unique source + 3 unique destination IPs
         assert '192.168.1.100' in unique_ips
         assert '192.168.1.101' in unique_ips
         assert '10.0.0.1' in unique_ips
         assert '10.0.0.2' in unique_ips
-        assert '10.0.0.3' in unique_ips 
+        assert '10.0.0.3' in unique_ips
