@@ -5,24 +5,19 @@ by the existing test suites. This includes advanced error handling, complex work
 and integration edge cases.
 """
 
+from unittest.mock import AsyncMock, patch
+
 import pytest
 import pytest_asyncio
-import asyncio
-from unittest.mock import AsyncMock, MagicMock, patch
-from datetime import datetime, timedelta
-from typing import Dict, Any, List, Optional
 
-from src.elasticsearch_client import ElasticsearchClient
-from src.dshield_client import DShieldClient
-from src.data_processor import DataProcessor
 from src.context_injector import ContextInjector
-from src.campaign_analyzer import CampaignAnalyzer
-from src.user_config import UserConfigManager, get_user_config, reset_user_config
+from src.data_processor import DataProcessor
+from src.user_config import get_user_config, reset_user_config
 
 
 class TestRemainingIntegration:
     """Test remaining integration scenarios and edge cases."""
-    
+
     @pytest_asyncio.fixture
     async def mock_es_client(self):
         """Create a mock ElasticsearchClient."""
@@ -34,10 +29,12 @@ class TestRemainingIntegration:
             client.get_available_indices = AsyncMock(return_value=["dshield-2024.01.01"])
             client.query_dshield_events = AsyncMock(return_value=([], 0, {}))
             client.stream_dshield_events = AsyncMock(return_value=([], 0, None))
-            client.stream_dshield_events_with_session_context = AsyncMock(return_value=([], 0, None, {}))
-            
+            client.stream_dshield_events_with_session_context = AsyncMock(
+                return_value=([], 0, None, {})
+            )
+
             yield client
-    
+
     @pytest_asyncio.fixture
     async def mock_dshield_client(self):
         """Create a mock DShieldClient."""
@@ -49,9 +46,9 @@ class TestRemainingIntegration:
             client.get_ip_reputation = AsyncMock(return_value={})
             client.get_top_attackers = AsyncMock(return_value=[])
             client.get_attack_summary = AsyncMock(return_value={})
-            
+
             yield client
-    
+
     @pytest.mark.asyncio
     async def test_complex_workflow_integration(self, mock_es_client, mock_dshield_client):
         """Test complex workflow integration with multiple components."""
@@ -63,76 +60,78 @@ class TestRemainingIntegration:
                 "source_ip": "192.168.1.100",
                 "destination_ip": "10.0.0.1",
                 "event_type": "network_scan",
-                "severity": "high"
+                "severity": "high",
             }
         ]
         mock_es_client.query_dshield_events.return_value = (mock_events, 1, {})
-        
+
         mock_reputation = {
             "ip_address": "192.168.1.100",
             "reputation_score": 85,
-            "threat_level": "high"
+            "threat_level": "high",
         }
         mock_dshield_client.get_ip_reputation.return_value = mock_reputation
-        
+
         # Test complete workflow
         data_processor = DataProcessor()
         context_injector = ContextInjector()
-        
+
         # Process events
         processed_events = data_processor.process_security_events(mock_events)
         assert len(processed_events) == 1, "Should process 1 event"
-        
+
         # Generate summary
         summary = data_processor.generate_security_summary(processed_events)
         assert summary["total_events"] == 1, "Should have 1 total event"
-        
+
         # Extract IPs for enrichment
         unique_ips = data_processor.extract_unique_ips(processed_events)
         assert "192.168.1.100" in unique_ips, "Should extract source IP"
-        
+
         # Mock threat intelligence
         threat_intel = {"192.168.1.100": mock_reputation}
-        
+
         # Prepare context
         context = context_injector.prepare_security_context(processed_events, threat_intel)
         assert "data" in context, "Should have data in context"
         assert "events" in context["data"], "Should have events in context data"
-        assert "threat_intelligence" in context["data"], "Should have threat intelligence in context data"
-        
+        assert "threat_intelligence" in context["data"], (
+            "Should have threat intelligence in context data"
+        )
+
         # Format for ChatGPT
         chatgpt_context = context_injector.inject_context_for_chatgpt(context)
         assert isinstance(chatgpt_context, str), "Should return formatted string"
         assert "192.168.1.100" in chatgpt_context, "Should include IP in context"
-    
+
     @pytest.mark.asyncio
     async def test_advanced_error_handling(self, mock_es_client):
         """Test advanced error handling scenarios."""
         # Test connection failure
         mock_es_client.connect.side_effect = Exception("Connection failed")
-        
+
         with pytest.raises(Exception, match="Connection failed"):
             await mock_es_client.connect()
-        
+
         # Test query failure
         mock_es_client.connect.side_effect = None
         mock_es_client.query_dshield_events.side_effect = Exception("Query failed")
-        
+
         with pytest.raises(Exception, match="Query failed"):
             await mock_es_client.query_dshield_events(time_range_hours=24)
-        
+
         # Test partial failure recovery
         mock_es_client.query_dshield_events.side_effect = None
         mock_es_client.query_dshield_events.return_value = ([], 0, {})
-        
+
         events, total_count, pagination_info = await mock_es_client.query_dshield_events(
             time_range_hours=24
         )
-        
+
         assert len(events) == 0, "Should handle empty results gracefully"
         assert total_count == 0, "Should have 0 total count"
         assert isinstance(pagination_info, dict), "Should return pagination info"
-    
+
     @pytest.mark.asyncio
     async def test_campaign_analysis_edge_cases(self, mock_es_client):
         """Test campaign analysis edge cases."""
@@ -141,124 +140,130 @@ class TestRemainingIntegration:
             analyzer = mock_class.return_value
             analyzer.elastic_client = mock_es_client
             analyzer.detect_campaigns = AsyncMock(return_value=[])
-            
+
             # Test with no events
             mock_es_client.query_dshield_events.return_value = ([], 0, {})
-            
+
             # Test campaign detection with empty data
             campaigns = await analyzer.detect_campaigns([])
             assert len(campaigns) == 0, "Should handle empty events"
-            
+
             # Test with single event
             mock_event = {
                 "id": "1",
                 "@timestamp": "2024-01-01T10:00:00Z",
-                "source_ip": "192.168.1.100"
+                "source_ip": "192.168.1.100",
             }
             mock_es_client.query_dshield_events.return_value = ([mock_event], 1, {})
-            
+
             campaigns = await analyzer.detect_campaigns([mock_event])
             assert isinstance(campaigns, list), "Should return list of campaigns"
-    
+
     @pytest.mark.asyncio
     async def test_user_configuration_edge_cases(self):
         """Test user configuration edge cases."""
         # Reset user config to ensure clean state
         reset_user_config()
         config_manager = get_user_config()
-        
+
         # Test invalid setting updates - these should be handled gracefully
         # The actual implementation may not raise ValueError for these cases
         try:
             config_manager.update_setting("query", "default_page_size", -1)
         except (ValueError, TypeError):
             pass  # Expected behavior
-        
+
         try:
             config_manager.update_setting("query", "max_page_size", 0)
         except (ValueError, TypeError):
             pass  # Expected behavior
-        
+
         try:
             config_manager.update_setting("query", "fallback_strategy", "invalid_strategy")
         except (ValueError, TypeError):
             pass  # Expected behavior
-        
+
         # Reset user config to ensure clean state after invalid updates
         reset_user_config()
         config_manager = get_user_config()
-        
+
         # Test valid setting updates - need to set max_page_size first
         config_manager.update_setting("query", "max_page_size", 1000)
         config_manager.update_setting("query", "default_page_size", 50)
         config_manager.update_setting("query", "fallback_strategy", "aggregate")
-        
+
         # Verify updates
         assert config_manager.get_setting("query", "default_page_size") == 50
         assert config_manager.get_setting("query", "max_page_size") == 1000
         assert config_manager.get_setting("query", "fallback_strategy") == "aggregate"
-        
+
         # Test configuration export
         config_data = config_manager.export_config()
         assert "query" in config_data, "Should export query settings"
         assert "pagination" in config_data, "Should export pagination settings"
         assert "streaming" in config_data, "Should export streaming settings"
-        
+
         # Test environment variables export
         env_vars = config_manager.get_environment_variables()
         assert "DEFAULT_PAGE_SIZE" in env_vars, "Should export DEFAULT_PAGE_SIZE"
         assert "MAX_PAGE_SIZE" in env_vars, "Should export MAX_PAGE_SIZE"
         assert "ENABLE_SMART_OPTIMIZATION" in env_vars, "Should export ENABLE_SMART_OPTIMIZATION"
-    
+
     @pytest.mark.asyncio
     async def test_streaming_edge_cases(self, mock_es_client):
         """Test streaming edge cases."""
         # Test streaming with invalid stream_id
         mock_es_client.stream_dshield_events.return_value = ([], 0, None)
-        
+
         events, total_count, stream_id = await mock_es_client.stream_dshield_events(
-            time_range_hours=24,
-            stream_id="invalid_stream_id"
+            time_range_hours=24, stream_id="invalid_stream_id"
         )
-        
+
         assert len(events) == 0, "Should handle invalid stream_id gracefully"
         assert stream_id is None, "Should return None for invalid stream_id"
-        
+
         # Test session streaming with empty session fields
         mock_session_context = {
             "session_fields": [],
             "sessions_in_chunk": 0,
             "session_summaries": [],
-            "performance_metrics": {"query_time_ms": 0}
+            "performance_metrics": {"query_time_ms": 0},
         }
         mock_es_client.stream_dshield_events_with_session_context.return_value = (
-            [], 0, None, mock_session_context
+            [],
+            0,
+            None,
+            mock_session_context,
         )
-        
-        events, total_count, next_stream_id, session_context = await mock_es_client.stream_dshield_events_with_session_context(
-            time_range_hours=24,
-            session_fields=[]
+
+        (
+            events,
+            total_count,
+            next_stream_id,
+            session_context,
+        ) = await mock_es_client.stream_dshield_events_with_session_context(
+            time_range_hours=24, session_fields=[]
         )
-        
+
         assert len(events) == 0, "Should handle empty session fields"
         assert session_context["sessions_in_chunk"] == 0, "Should have 0 sessions"
-    
+
     @pytest.mark.asyncio
     async def test_data_processing_edge_cases(self):
         """Test data processing edge cases."""
         data_processor = DataProcessor()
-        
+
         # Test with None events - should handle gracefully
         try:
             processed_events = data_processor.process_security_events(None)
             assert len(processed_events) == 0, "Should handle None events"
         except (TypeError, AttributeError):
             pass  # Expected behavior for None input
-        
+
         # Test with empty events
         processed_events = data_processor.process_security_events([])
         assert len(processed_events) == 0, "Should handle empty events"
-        
+
         # Test with events containing None values
         events_with_none = [
             {
@@ -266,42 +271,44 @@ class TestRemainingIntegration:
                 "timestamp": None,
                 "source_ip": None,
                 "destination_ip": None,
-                "event_type": None
+                "event_type": None,
             }
         ]
         processed_events = data_processor.process_security_events(events_with_none)
         assert len(processed_events) == 1, "Should process events with None values"
-        
+
         # Test summary generation with empty events
         summary = data_processor.generate_security_summary([])
         assert summary["total_events"] == 0, "Should handle empty events in summary"
         assert summary["unique_source_ips"] == 0, "Should have 0 unique source IPs"
         assert summary["unique_destination_ips"] == 0, "Should have 0 unique destination IPs"
-        
+
         # Test IP extraction with empty events
         unique_ips = data_processor.extract_unique_ips([])
         assert len(unique_ips) == 0, "Should handle empty events in IP extraction"
-    
+
     @pytest.mark.asyncio
     async def test_context_injection_edge_cases(self):
         """Test context injection edge cases."""
         context_injector = ContextInjector()
-        
+
         # Test with None data - should handle gracefully
         try:
             context = context_injector.prepare_security_context(None, None)
             assert "data" in context, "Should handle None events"
             assert "events" in context["data"], "Should handle None events"
-            assert "threat_intelligence" in context["data"], "Should handle None threat intelligence"
+            assert "threat_intelligence" in context["data"], (
+                "Should handle None threat intelligence"
+            )
         except (TypeError, AttributeError):
             pass  # Expected behavior for None input
-        
+
         # Test with empty data
         context = context_injector.prepare_security_context([], {})
         assert "data" in context, "Should handle empty events"
         assert "events" in context["data"], "Should handle empty events"
         # Note: threat_intelligence may not be present in empty context
-        
+
         # Test ChatGPT formatting with empty context
         try:
             chatgpt_context = context_injector.inject_context_for_chatgpt({})
@@ -310,80 +317,74 @@ class TestRemainingIntegration:
         except KeyError:
             # Expected behavior for empty context without required keys
             pass
-        
+
         # Test with malformed data
         malformed_events = [{"invalid": "data"}]
         context = context_injector.prepare_security_context(malformed_events, {})
         assert "data" in context, "Should handle malformed events"
         assert "events" in context["data"], "Should have events in context data"
-        
+
         chatgpt_context = context_injector.inject_context_for_chatgpt(context)
         assert isinstance(chatgpt_context, str), "Should handle malformed data"
-    
+
     @pytest.mark.asyncio
     async def test_performance_optimization_edge_cases(self, mock_es_client):
         """Test performance optimization edge cases."""
         # Test with very large page size
         mock_es_client.query_dshield_events.return_value = ([], 10000, {})
-        
+
         events, total_count, pagination_info = await mock_es_client.query_dshield_events(
-            time_range_hours=24,
-            page_size=10000
+            time_range_hours=24, page_size=10000
         )
-        
+
         assert total_count == 10000, "Should handle large page sizes"
-        
+
         # Test with very small page size
         events, total_count, pagination_info = await mock_es_client.query_dshield_events(
-            time_range_hours=24,
-            page_size=1
+            time_range_hours=24, page_size=1
         )
-        
+
         assert total_count == 10000, "Should handle small page sizes"
-        
+
         # Test with very long time range
         events, total_count, pagination_info = await mock_es_client.query_dshield_events(
             time_range_hours=8760,  # 1 year
-            page_size=100
+            page_size=100,
         )
-        
+
         assert total_count == 10000, "Should handle long time ranges"
-    
+
     @pytest.mark.asyncio
     async def test_integration_error_recovery(self, mock_es_client, mock_dshield_client):
         """Test integration error recovery scenarios."""
         # Test partial failure in workflow
         mock_events = [
-            {
-                "id": "1",
-                "@timestamp": "2024-01-01T10:00:00Z",
-                "source_ip": "192.168.1.100"
-            }
+            {"id": "1", "@timestamp": "2024-01-01T10:00:00Z", "source_ip": "192.168.1.100"}
         ]
-        
+
         # Mock successful event query but failed enrichment
         mock_es_client.query_dshield_events.return_value = (mock_events, 1, {})
         mock_dshield_client.get_ip_reputation.side_effect = Exception("Enrichment failed")
-        
+
         # Test that the system can continue without enrichment
         data_processor = DataProcessor()
         processed_events = data_processor.process_security_events(mock_events)
-        
+
         assert len(processed_events) == 1, "Should process events even without enrichment"
-        
+
         # Test context injection with failed enrichment
         context_injector = ContextInjector()
         context = context_injector.prepare_security_context(processed_events, {})
-        
+
         assert "data" in context, "Should create context even without threat intelligence"
         assert "events" in context["data"], "Should have events in context"
         # Note: threat_intelligence may not be present in context without threat intel
-        
+
         # Test ChatGPT formatting with partial data
         chatgpt_context = context_injector.inject_context_for_chatgpt(context)
         assert isinstance(chatgpt_context, str), "Should format partial context"
         # Note: IP may not be directly visible in formatted output, but context should be valid
-    
+
     @pytest.mark.asyncio
     async def test_comprehensive_workflow_validation(self, mock_es_client, mock_dshield_client):
         """Test comprehensive workflow validation."""
@@ -396,7 +397,7 @@ class TestRemainingIntegration:
                 "destination_ip": "10.0.0.1",
                 "event_type": "network_scan",
                 "severity": "high",
-                "category": "intrusion_detection"
+                "category": "intrusion_detection",
             },
             {
                 "id": "2",
@@ -405,30 +406,30 @@ class TestRemainingIntegration:
                 "destination_ip": "10.0.0.2",
                 "event_type": "authentication_failure",
                 "severity": "medium",
-                "category": "authentication"
-            }
+                "category": "authentication",
+            },
         ]
-        
+
         mock_es_client.query_dshield_events.return_value = (mock_events, 2, {})
-        
+
         mock_reputation = {
             "ip_address": "192.168.1.100",
             "reputation_score": 85,
             "threat_level": "high",
             "attack_count": 150,
             "first_seen": "2024-01-01T00:00:00Z",
-            "last_seen": "2024-01-01T23:59:59Z"
+            "last_seen": "2024-01-01T23:59:59Z",
         }
         mock_dshield_client.get_ip_reputation.return_value = mock_reputation
-        
+
         # Execute complete workflow
         data_processor = DataProcessor()
         context_injector = ContextInjector()
-        
+
         # Process events
         processed_events = data_processor.process_security_events(mock_events)
         assert len(processed_events) == 2, "Should process 2 events"
-        
+
         # Generate summary
         summary = data_processor.generate_security_summary(processed_events)
         assert summary["total_events"] == 2, "Should have 2 total events"
@@ -439,31 +440,33 @@ class TestRemainingIntegration:
             assert summary["high_severity_events"] == 1, "Should have 1 high severity event"
         if "medium_severity_events" in summary:
             assert summary["medium_severity_events"] == 1, "Should have 1 medium severity event"
-        
+
         # Extract IPs
         unique_ips = data_processor.extract_unique_ips(processed_events)
         assert "192.168.1.100" in unique_ips, "Should extract source IP"
         assert "10.0.0.1" in unique_ips, "Should extract destination IP 1"
         assert "10.0.0.2" in unique_ips, "Should extract destination IP 2"
-        
+
         # Mock threat intelligence
         threat_intel = {"192.168.1.100": mock_reputation}
-        
+
         # Prepare context
         context = context_injector.prepare_security_context(processed_events, threat_intel)
         assert "data" in context, "Should have data in context"
         assert "events" in context["data"], "Should have events in context data"
-        assert "threat_intelligence" in context["data"], "Should have threat intelligence in context data"
+        assert "threat_intelligence" in context["data"], (
+            "Should have threat intelligence in context data"
+        )
         assert len(context["data"]["events"]) == 2, "Should have 2 events in context"
         # Check that threat intelligence contains the IP in the expected structure
         threat_intel = context["data"]["threat_intelligence"]
         assert "high_risk_ips" in threat_intel, "Should have high_risk_ips in threat intelligence"
         high_risk_ips = [ip["ip"] for ip in threat_intel["high_risk_ips"]]
         assert "192.168.1.100" in high_risk_ips, "Should have threat intelligence for IP"
-        
+
         # Format for ChatGPT
         chatgpt_context = context_injector.inject_context_for_chatgpt(context)
         assert isinstance(chatgpt_context, str), "Should return formatted string"
         assert "192.168.1.100" in chatgpt_context, "Should include source IP"
         # Note: Destination IPs and event types may not be directly visible in formatted output
-        # but the context should be valid and contain the essential information 
+        # but the context should be valid and contain the essential information
